Napkin.Auth.GoogleOAuth

Documentation

initOAuthConf :: OAuth2ClientId -> OAuth2ClientSecret -> OAuth2Callback -> OAuth2Conf

initialize napkin oAuth configuration Napkin takes clientId, clientSecret and callback url to initiazlie the oauth configs. configuation parameters are from google oauth api docs, https://developers.google.com/identity/protocols/oauth2/native-app#create-code-challenge

initOAuthRunTime :: MonadIO m => m OAuth2RunTime

initialize napkin oAuth runtime parameters

configureGoogleOAuthEnv :: (MonadReader OAuth2 m, MonadIO m) => m OAuth2Env

getOAuth2Url :: OAuth2Env -> Text

createTmpCredentialFile :: (MonadIO m, MonadCatch m, Katip m) => OAuth2Env -> m FilePath

creates a temporary credential file We attemtp to create a temp credential file. if we can create a Bigquery google environment, then the tmp file is cashed for future use

authWorkflow :: (MonadIO m, MonadCatch m, Katip m) => OAuth2Env -> m ()

napkin authentication workflow The work-flow : + creates credential file and stores in OS $TMP folder + attempts to create google env and bigquery env + if success, cache the tmp file napkin well know location for future use.

authResetWorkflow :: (MonadIO m, MonadCatch m, Katip m) => OAuth2Env -> m ()

napkin reset authentication workflow removes the previously cached napkin auth file for this project

makeGoogleEnvWithCredentials :: (MonadIO m, MonadCatch m, Katip m) => Credentials BigQueryScopes -> m (Env BigQueryScopes)

makeGoogleEnv

Arguments

:: (MonadCatch m, Katip m)

=> Maybe FilePath
Optional location of credentials file. Defaults to ~.configgcloud/application_default_credentials.json

-> m GoogleEnv

authorizationUri :: URI

construct the google oauth endpoint for the OAuth2 Googlew APIs access for additional detail see https://developers.google.com/identity/protocols/oauth2#installed

napkinCredFile :: IO FilePath

appendQueryParams :: URIQueryParams -> URIRef a -> URIRef a

toGCredential :: OAuth2Conf -> OAuth2Token -> GCredentials

xchgCodeForToken :: (MonadIO m, MonadCatch m) => OAuth2Env -> OAuth2Code -> m (Either AuthEvalError OAuth2Token)