Napkin.Render.Types

Documentation

sqlToFile :: RenderSql a b => b -> a -> FilePath -> IO ()

Render to a local SQL file, mainly for debugging/devel help.

sqlToFileCollapsed :: RenderSql a b => b -> a -> FilePath -> IO ()

printSql :: RenderSql a b => b -> a -> IO ()

renderSql' :: RenderSql a b => b -> a -> String

renderSqlWith :: RenderSql a b => Style -> b -> a -> String

renderSqlCollapsed :: RenderSql a b => b -> a -> String

styleCollapsed :: Style

class RenderSql a backend where

If one day we support multiple backends, one option would be for this typeclass to take a (Proxy backend) argument to dispatch on the right backend.

Methods

renderSql :: backend -> a -> Doc

class CommandDescr a backend where

Methods

commandDescr :: backend -> a -> String

data RenderedSql

Constructors

forall a backend.RenderSql a backend => RenderedSql backend a

unpackRenderedSql :: RenderedSql -> String

class ReplaceReference backend where

Ability to hackishly replace references to a given table within a SQL string blob. Enables namespacing and we wouldn't need it if we could comprehensively parse each backend's SQL.

Minimal complete definition

Nothing

Methods

replaceReference

Arguments

:: Proxy backend

-> Ref Table
Search for this table

-> Ref Table
Replace with this table

-> Text
I give you a whole SQL blob fixer

-> Text