Napkin.Run.Assertions

Documentation

assertRowCountsWithin

Arguments

:: (Katip m, MonadFail m, Backend b, MonadQueryStats m, RenderSql Query b)

=> BackendConn b

-> String
A message to tag test with

-> Double
Tolerance, 0-1

-> Relation
Expected relation

-> Relation
Relation being checked

-> m ()

assertAnswersWithin :: (Katip m, MonadFail m, Backend b, RenderSql q1 b, RenderSql q2 b, MonadQueryStats m) => BackendConn b -> String -> Double -> q1 -> q2 -> m ()

todo: how to make this work both for ints and doubles?

assertUniqueBy :: (RenderSql Query b, Backend b, Katip m, MonadFail m, MonadQueryStats m) => Relation -> [Ref a] -> String -> BackendConn b -> m ()

assertUniquesBy

Arguments

:: (Katip m, MonadFail m, Backend b, RenderSql Query b, MonadQueryStats m)

=> Ref a1

-> [(SExp, [Ref a])]
A set of subcases that display uniqueness

-> String

-> BackendConn b

-> m ()

Like assertUniqeBy but can express multiple cases on same table

assertUniqueByCap :: (Katip m, MonadFail m, Backend b, RenderSql Query b, MonadQueryStats m) => Double -> Relation -> [Ref a] -> String -> BackendConn b -> m ()