Napkin.Run.BigQuery

Documentation

data BigQueryRunError

Constructors

ProtoErrors !(NonEmpty ErrorProto)

MissingDatasetId

MissingProjectId

mkBigQueryEnv

Arguments

:: (MonadCatch m, Katip m)

=> Maybe FilePath
Optional location of credentials file. Defaults to ~.configgcloud/application_default_credentials.json

-> BQProjectId
Default project

-> Maybe BQDatasetId
Default dataset

-> HashMap Text Text

-> m (BackendConn BigQuery)

mkBigQueryEnv'

Arguments

:: (MonadCatch m, MonadIO m)

=> IO GoogleEnv
How do we set up each entry in the GoogleEnv pool?

-> BQProjectId
Default project

-> Maybe BQDatasetId
Default dataset

-> HashMap Text Text

-> m (BackendConn BigQuery)

Make a BigQuery env but with control over configuring how the Google environment is set up. You should usually call mkGoogleEnv and then fmap some customization over it.

data family BackendConn b

Utilities

exportTable

Arguments

:: (Katip m, MonadQueryStats m)

=> Pool GoogleEnv

-> BQProjectId

-> BQDatasetId

-> BQTableId

-> GCSBucket
Google storage destination bucket

-> GCSKey
Google storage destination key

-> m Job

downloadGCSFile

Arguments

:: (MonadIO m, MonadUnliftIO m)

=> Pool GoogleEnv

-> GCSBucket

-> GCSKey

-> FilePath
Destination

-> m ()

uploadFileToGCS

Arguments

:: (MonadIO m, MonadUnliftIO m)

=> Pool GoogleEnv

-> GCSBucket

-> GCSKey

-> FilePath
Source

-> Maybe MediaType
Defaults to application/octet-stream. note MediaType has an IsString instance

-> m ()

getGCSFilePermissions :: MonadIO m => Pool GoogleEnv -> GCSBucket -> GCSKey -> m [(GCSACLEntity, GCSACLRole)]

setGCSPermissionsForEntity :: MonadUnliftIO m => Pool GoogleEnv -> GCSBucket -> GCSKey -> GCSACLEntity -> GCSACLRole -> m ()

decomposeRef :: Ref Table -> (Maybe BQProjectId, Maybe BQDatasetId, BQTableId)

segs are [project[dataset]]table. Usually combines with default dataset id and project id to fully resolve Refs.

Orphan instances

Backend BigQuery

Instance details

Associated Types

data BackendConn BigQuery

Methods

backendQuery :: (Katip m, MonadQueryStats m, RenderSql q BigQuery) => BackendConn BigQuery -> q -> m [Map Text Value]

backendExecute :: (Katip m, MonadQueryStats m, RenderSql q BigQuery) => BackendConn BigQuery -> q -> m ()

Command Grant () BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> Grant -> m ()

Command CheckTableExists Bool BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> CheckTableExists -> m Bool

Command DeleteFrom Job BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> DeleteFrom -> m Job

Command InsertIntoQuery Job BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> InsertIntoQuery -> m Job

Command DropTable () BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> DropTable -> m ()

Command DropView () BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> DropView -> m ()

Command CreateViewAs Table BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> CreateViewAs -> m Table

Command CopyTable () BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> CopyTable -> m ()

Command RenameTable () BigQuery
BQ implements rename via copy-then-drop

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> RenameTable -> m ()

Command GetRelationSchema (Fields Field) BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> GetRelationSchema -> m (Fields Field)

Command (CreateTableAs ()) Job BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> CreateTableAs () -> m Job

Command (CreateTableAs TableMeta) Job BigQuery

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn BigQuery -> CreateTableAs TableMeta -> m Job