Napkin.Run.Composite

Composite commands

recreate :: (Drop a c, Katip m, Command c t1 b, Command a t2 b, MonadQueryStats m) => BackendConn b -> a -> m ()

Drop table if exists and re-create

recreateView :: (Katip m, TableRef c, Command DropView t1 b, Command c t2 b, MonadQueryStats m) => BackendConn b -> c -> m ()

recreateSwap :: (Command RenameTable a backend, Command create a2 backend, Command drop a1 backend, Drop create drop, TableRef create, MonadQueryStats f, Katip f) => BackendConn backend -> create -> (Ref Table -> f ()) -> f ()

incrementalInsert

Arguments

:: (Katip m, Command CheckTableExists Bool b, Command (CreateTableAs meta) t1 b, Command InsertIntoQuery t1 b, Command DeleteFrom t2 b, MonadQueryStats m, Command DropTable t3 b)

=> BackendConn b

-> (Day -> IO (CreateTableAs meta))
Table definition given a predicate expression. Must use given Day as a cutoff

-> (Day -> SExp)
How to delete rows after a given date in the destination table

-> Day
Default starting point date during fresh creation

-> Integer
# of days to erase in increment

-> m ()

Either create the table or incrementally add to it based on a timestamp column.

mkDeleteCutoff :: Ref SExp -> Day -> SExp

Make a cutoff predicated from a column name for a date

readSqlFileWithCutoff :: forall b m. (RenderSql SExp b, MonadIO m, ParseSQL b) => b -> FilePath -> SqlTemplateVariables -> Day -> m Query

Read local sql file with a {{cutoff}} mustache variable available

incrementalInsertByPK

Arguments

:: (Command InsertIntoQuery a backend, Command CheckTableExists Bool backend, Command (CreateTableAs m1) a backend, MonadQueryStats m, Katip m)

=> BackendConn backend

-> (SExp -> CreateTableAs m1)

-> SExp
Primary key column

-> SExp
Starting timestamp value in fresh create

-> Integer
Num days to look back in each update

-> m ()

ctaToIncrementalInsertByPK :: MonadIO m => (SExp -> CreateTableAs m1) -> SExp -> Integer -> m InsertIntoQuery

Turn a create query into an incremental insert into query.