Napkin.Run.Redshift

Redshift Re-exports

data RedshiftT (m :: Type -> Type) a

The Redshift monad transformer that all the redshift API operations run in. This monad ensures that all tunnels get a unique port on localhost.

connectRedshift :: forall (m :: Type -> Type). MonadIO m => RedshiftDB -> Int -> NominalDiffTime -> Int -> RedshiftT m (Pool Connection)

Establish a postgresql-simple connection to a RedshiftDB.

runRedshift :: MonadIO m => RedshiftState -> RedshiftT m a -> m a

Runs any DB action in IO.

The only way to unwrap RedshiftT, as we need to carefully manage SSH tunnels and database connections.

defRS :: Configuration -> RedshiftState

A default RedshiftState that cleans up and is not verbose. If you store your keys in ~/.aws-keys you can use baseConfiguration for the configuration parameter

redNick :: String -> IO RedshiftDB

Retrieves a RedshiftDB for the specified DB nickname.

withRedshift :: String -> Int -> NominalDiffTime -> Int -> (BackendConn Redshift -> IO b) -> IO b

withPostgresString :: String -> (BackendConn Redshift -> IO b) -> IO b

Orphan instances

Backend Redshift

Instance details

Associated Types

data BackendConn Redshift

Methods

backendQuery :: (Katip m, MonadQueryStats m, RenderSql q Redshift) => BackendConn Redshift -> q -> m [Map Text Value]

backendExecute :: (Katip m, MonadQueryStats m, RenderSql q Redshift) => BackendConn Redshift -> q -> m ()

HasPostgresConn Redshift

Instance details

Methods

getPostgresConn :: BackendConn Redshift -> Pool Connection

Command Grant Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> Grant -> m Int64

Command CheckTableExists Bool Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> CheckTableExists -> m Bool

Command DeleteFrom Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> DeleteFrom -> m Int64

Command InsertIntoQuery Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> InsertIntoQuery -> m Int64

Command DropTable Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> DropTable -> m Int64

Command DropView Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> DropView -> m Int64

Command CreateViewAs Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> CreateViewAs -> m Int64

Command RenameTable Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> RenameTable -> m Int64

Command GetRelationSchema (Fields Field) Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> GetRelationSchema -> m (Fields Field)

Command (CreateTableAs ()) Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> CreateTableAs () -> m Int64

Command (CreateTableAs TableMeta) Int64 Redshift

Instance details

Methods

execCommand :: (Katip m, MonadQueryStats m) => BackendConn Redshift -> CreateTableAs TableMeta -> m Int64