Napkin.Spec

Documentation

type Distinct = Bool

data HookAssertionError

Constructors

HookAssertionError [Relation] String

type CustomValidator = SpecMetaArgs -> Ref Table -> Maybe FilePath -> Query -> [String]

CustomValidators are run before namespacing and return a list of errors in the incoming Spec.

data ValidationError

Constructors

CycleDetectedError [[Ref Table]]

MissingSQLFileError (Ref Table) FilePath

SQLParseError String (Ref Table)

UndeclaredDependenciesError (Ref Table) [Ref Table]

ExtraneousDependenciesError (Ref Table) [Ref Table]

CustomValidatorError String

CustomPreprocessorError String

YamlError String

HaskellInterpreterError InterpreterError

type IsInteractive = Bool

data CmdError

Constructors

CmdError

Fields

cmdError_code :: Int

Exit code

cmdError_stderrExcerpt :: Text

data CmdFailureBehavior

How should cmd behave if the command exits with a failure

Constructors

ThrowOnFailure
Log the process' stderr and throw an exception, interrupting the spec

LogOnFailure
Log the process' stderr but do not throw an exception, allowing the spec to continue

type DependencySet = Set (Ref Table)

type SpecBackend bk drop t1 t2 t3 t4 t5 t6 t7 = (Command CreateViewAs t3 bk, Command CheckTableExists Bool bk, Command Grant t1 bk, Command (CreateTableAs (BackendTableMeta bk)) t4 bk, Command InsertIntoQuery t6 bk, Command drop t5 bk, Command DropView t2 bk, Command RenameTable t7 bk, Drop (CreateTableAs (BackendTableMeta bk)) drop, Default (BackendTableMeta bk), RenderSql (Ref Table) bk, RenderSql Query bk, RenderSql SExp bk, ReplaceReference bk)

What we seem to need to be able to call executeSpec; it's basically all about choosing a backend that satisfies the needed operations.

newtype SpecOpts

Constructors

SpecOpts

Fields

_spDry :: Bool

data SpecRuntime b

Constructors

SpecRuntime

Fields

_spPipeline :: Pipeline

_spConn :: BackendConn b

spec :: Ref Table -> Maybe (BackendTableMeta b) -> TableSpec b

shorthand for starting specs, use lenses to then update parts

ctaToSpec :: CreateTableAs (BackendTableMeta b) -> TableSpec b

cvaToSpec :: CreateViewAs -> TableSpec b

splitIfUnion :: TableSpec b0 -> TableSpec b0

Split UNION queries and insert each individually.

splitIfUnionBy :: Int -> TableSpec b0 -> TableSpec b0

Split a top level UNION based query table definition into individual insert-based queries in chunks of n.

specIncrementalByPK

Arguments

:: (Command CheckTableExists Bool b0, Command InsertIntoQuery a b0, Command (CreateTableAs (BackendTableMeta b0)) a b0)

=> (SExp -> CreateTableAs (BackendTableMeta b0))
Make me a query given a date cutoff

-> SExp
Expression that produces the identity/primary key for this table

-> SExp
A start date if the table needs to be freshly created

-> Integer
# of days to go back and refresh each time update is run

-> TableSpec b0

specIncrementalByTime :: (MonadIO m, Command (CreateTableAs (BackendTableMeta b)) t1 b, Command CheckTableExists Bool b, Command InsertIntoQuery t1 b, Command DeleteFrom t2 b, Command DropTable t3 b, Command RenameTable t4 b, Drop (CreateTableAs (BackendTableMeta b)) DropTable) => (Day -> IO (CreateTableAs (BackendTableMeta b))) -> (Day -> SExp) -> Day -> Integer -> [Ref Table] -> m (TableSpec b)

specIncrementalByTimeCreateAction :: (MonadIO m, Katip m, MonadQueryStats m, Command (CreateTableAs (BackendTableMeta b)) t1 b, Command CheckTableExists Bool b, Command InsertIntoQuery t1 b, Command DeleteFrom t2 b, Command DropTable t3 b, Command RenameTable t4 b, Drop (CreateTableAs (BackendTableMeta b)) DropTable) => (Day -> IO (CreateTableAs (BackendTableMeta b))) -> Day -> (Day -> SExp) -> Integer -> TableStatus -> QueryTransformer -> SpecMetaArgs -> BackendConn b -> m ()

specIncrementalByTimeRawSql

Arguments

:: (MonadIO m, RenderSql SExp b, Default (BackendTableMeta b), Command (CreateTableAs (BackendTableMeta b)) t1 b, Command CheckTableExists Bool b, Command InsertIntoQuery t1 b, Command DeleteFrom t2 b, Command DropTable t3 b, Command RenameTable t4 b, Drop (CreateTableAs (BackendTableMeta b)) DropTable, ParseSQL b)

=> b
Name of the backend

-> Ref Table
Name of the target table

-> Maybe (BackendTableMeta b)
Any backend specific table options

-> [Ref Table]
Table's dependencies

-> FilePath
Local SQL file

-> SqlTemplateVariables
Any additional variables you want to bind

-> Ref SExp
Timestamp column in the destination table

-> Day
Start time when creating table fresh

-> Integer
N days to go bac

-> m (TableSpec b)

rawSqlSpec :: Ref Table -> Maybe (BackendTableMeta b) -> FilePath -> SpecTarget -> [Ref Table] -> Spec b (TableSpec b)

Read a file that contains raw SQL query code that can be used to instantiate a table.

rawSqlMustacheSpec :: Ref Table -> Maybe (BackendTableMeta b) -> FilePath -> SqlTemplateVariables -> SpecTarget -> [Ref Table] -> Spec b (TableSpec b)

Same as rawSqlSpec but allows for mustache templating inside of the raw SQL file.

longToWideAutoSpec

Arguments

:: (Backend b, SpecBackend b drop t1 t2 t3 t4 t5 t6 t7)

=> Ref Table
the new table to create

-> Ref Table
the long table from which this table will be derived- make sure it is also listed in the dependencies

-> [Alias SExp]
the key columns which will remain in the resultant wide version as-is

-> Ref SExp
the column name in the long table which will be split into multiple columns in the wide version

-> Ref SExp
the column in the long table which will be emitted from matching on the new columns

-> (Value -> Ref SExp)
a function to generate column names from the values found in the database

-> Maybe (BackendTableMeta b)
standard table options

-> SpecTarget
create a view or table

-> [Ref Table]
this table's dependencies (include the long table)

-> Spec b (TableSpec b)

Query the "long" version of the table to determine what columns to put in the "wide" version. This function does not support mustache templates.

simpleValueAsColumnName :: Ref SExp -> Value -> Ref SExp

A basic means of generating new column names from

commandSpec

Arguments

:: Command CheckTableExists Bool b

=> Ref Table
the name of the new table

-> String
command

-> [String]
command arguments

-> [Ref Table]
the list of dependencies- these cannot be deduced

-> Spec b (TableSpec b)

A custom Spec which runs a command processed with mustache with the metaargs dictionary.

combineTablesSpec

Arguments

:: (SpecBackend bk drop t1 t2 t3 t4 t5 t6 t7, Command GetRelationSchema (Fields Field) bk)

=> Ref Table

-> Maybe (BackendTableMeta bk)

-> SpecTarget

-> (Query -> Query)
Any final optional query modification

-> [Relation]
List of sources to combine

-> TableSpec bk

Specialized convenience for UNION-ing several tables and queries, even for cases where the list of fields available in each table is not known at compile time. This will introspect list of fields for cases where a bare table name is given instead of a QueryRelation.

defineTable :: TableSpec b -> Spec b ()

defineTableM :: Spec b (TableSpec b) -> Spec b ()

Ridiculous convenience so some users don't have to bind themselves

addHooks :: [([Ref Table], Hook b)] -> Spec b ()

addTags :: [TableSpecTag] -> Spec b () -> Spec b ()

checkUpdateStrategy :: TableSpec b -> TableStatus -> UpdateStrategy -> IO Bool

Should I be running the update?

diffDependencies :: Bool -> DependencySet -> DependencySet -> DependencySet

Return the set difference of the dependencies of second set from the first set, paying attention to BigQuery's * table name syntax.

asteriskRefMatch :: Ref Table -> Ref Table -> Bool

Return True if the asterisk Ref matches the second table's name.

asteriskExpand :: Ref Table -> DependencySet -> DependencySet

executeSpec :: forall m s bk drop t1 t2 t3 t4 t5 t6 t7. (MonadFix m, MonadBaseControl IO m, MonadMask m, Katip m, HasSpec s bk, ParseSQL bk, MonadReader ComboEnv m, SpecBackend bk drop t1 t2 t3 t4 t5 t6 t7, Backend bk, MonadQueryStats m) => bk -> SpecRuntime bk -> SpecOpts -> CustomValidator -> s -> m Success

dryRun :: (MonadIO m, Katip m) => Pipeline -> Ref Table -> DepChain -> t -> m (Either a (), SpecRun)

A run that succeeds immediately without ever actually running anything.

recreateStandard

Arguments

:: forall b t1 t2 t4 t5 t6 c f d. (Command CreateViewAs t2 b, Command DropView t1 b, Command (CreateTableAs d) t4 b, Command RenameTable t6 b, Drop (CreateTableAs d) c, Command c t5 b, Katip f, BackendTableMeta b ~ d, Default d, MonadQueryStats f)

=> BackendConn b

-> Ref Table

-> Maybe (BackendTableMeta b)

-> Query

-> SpecTarget

-> (Ref Table -> f ())
Callback to let you do additional things to the temporary table

-> f ()

dryExecute :: (LogItem a, Katip m) => t -> a -> m ()

cmd

Arguments

:: (Katip m, MonadThrow m)

=> CmdFailureBehavior

-> String
Command to run

-> [String]
Arguments

-> m ()

Combinator for use in things spec hooks to shell out to a command. The command being run will be logged, and on error the stderr will be trimmed and logged. When ThrowOnFailure is passed, a CmdError will be thrown.

reportFatalError :: Maybe RuntimeEventQueue -> IO ()

reportProgress :: Show a => Maybe RuntimeEventQueue -> Ref Table -> ComboT IO (Either a b, SpecRun) -> ComboT IO (Either a b, SpecRun)

Report runtime updates as the Spec execution progresses.

reportCompletion :: MonadIO m => Maybe RuntimeEventQueue -> TrackingQueryStats -> m ()

reportSkipUpdate :: MonadIO m => Maybe RuntimeEventQueue -> Ref Table -> m ()

emptyValidator :: CustomValidator

validateAndPreprocessSpecs :: (MonadIO m, ParseSQL bk) => bk -> CustomValidator -> Specs b -> m (Either [ValidationError] (Specs b))

Run basic checks against all Specs and convert every raw query to normal Query.

validateDependencies :: Ref Table -> Query -> [Ref Table] -> [Ref Table] -> ([ValidationError], [Ref Table])

diffRowCountsError :: (Relation, Integer) -> (Relation, Integer) -> HookAssertionError

greaterThanCountsError :: (Relation, Integer) -> (Relation, Integer) -> HookAssertionError

greaterThanConstantError :: (Relation, Integer) -> (Relation, Integer) -> HookAssertionError

uniquenessError :: (Relation, Integer) -> Integer -> SExp -> [Ref a] -> HookAssertionError

runHook :: forall backend. (Backend backend, RenderSql Query backend, RenderSql SExp backend) => BackendConn backend -> Hook backend -> ComboT IO ()

countCompare :: (Backend backend, RenderSql Query backend) => BackendConn backend -> Relation -> Relation -> (Integer -> Integer -> Bool) -> ((Relation, Integer) -> (Relation, Integer) -> HookAssertionError) -> ComboT IO ()

tableCount :: Distinct -> Relation -> Query

runRelationCount :: (Backend bk, RenderSql Query bk) => BackendConn bk -> Distinct -> Relation -> ComboT IO Integer

assertNotAllNull :: SExp -> Relation -> Hook b

prettyQueryStatsForTable :: Ref Table -> TrackingQueryStats -> String

processNamespaceQuery :: ParseSQL b => b -> QueryTransformer -> Query -> SqlSource -> IO Query

Process any raw queries within a Query with the renamer.

logFinalReport :: (Katip m, RenderSql (Ref Table) bk, MonadQueryStats m) => bk -> [(Ref Table, Either SomeException (Either b SpecRun, TrackingQueryStats))] -> [Either SomeException (Bool, TrackingQueryStats)] -> m ()

specTagRefMap :: [TableSpec b] -> Map (Maybe TableSpecTag) [Ref Table]

create a mapping of TableSpecTags to `Ref Table`s for dumping.

applyQueryTransformer :: QueryTransformer -> CreateAction b -> CreateAction b

blacklistDependencies :: [Ref Table] -> TableSpec b -> TableSpec b

module Napkin.Spec.Types

type MonadCombo m = (MonadFix m, MonadBaseControl IO m, MonadIO m, MonadMask m, MonadFail m)

newtype SqlTemplateVariables

Constructors

SqlTemplateVariables

Fields

_sqlTemplateVariables :: Map Text Value

sqlTemplate :: [(Text, Value)] -> SqlTemplateVariables

sqlVar :: ToJSON ω => Text -> ω -> (Text, Value)