Napkin.Spec.Cli

Description

Napkin Command Line Interface, CLI

Documentation

data CliOptions

Data structure for all CLI commnad and sub-commands.

SubCommand data strucres are Run, Dump, ListTags, DumpDeps Each SubCommands data strucures contain the available commands

Constructors

Run RunOptions

Dump DumpOptions

ListTags

DumpDeps

data CustomSpecPreprocessor bk

User-specified function to preprocess all specs. Useful for renaming all tables, for example, in development vs. production mode.

Constructors

CustomSpecPreprocessor (CustomSpecPreprocessorFunc bk)

type CustomSpecPreprocessorFunc bk = SpecMetaArgs -> [(SpecRuntime bk, Specs bk)] -> [(SpecRuntime bk, Specs bk)]

emptyPreprocessor :: CustomSpecPreprocessor bk

The default preprocessor which makes no changes to the specs.

data DumpOptions

Data structure dump sub-command.

Constructors

DumpOptions

Fields

_dumpOptions_outputDirectory :: FilePath

_dumpOptions_force :: Bool

_dumpOptions_generateInserts :: Bool

_dumpOptions_meta :: [ByteString]

_dumpOptions_excludeUnmanagedTables :: Bool

parseFilePath :: ReadM FilePath

parseMeta :: ReadM ByteString

data RunOptions

Data structure run sub-command.

Constructors

RunOptions

Fields

_runOptions_disableAll :: Bool

_runOptions_enableAll :: Bool

_runOptions_depsOnly :: Bool

_runOptions_forceTables :: [Ref Table]

_runOptions_skipTables :: [Ref Table]

_runOptions_naturalUpdates :: [Ref Table]

_runOptions_specSets :: [Pipeline]

_runOptions_app :: Text

_runOptions_logLevel :: Maybe Severity

_runOptions_dryRun :: Bool

_runOptions_namespace :: Maybe Text

_runOptions_sqlite :: Maybe FilePath

_runOptions_showProgress :: Bool

_runOptions_meta :: [ByteString]

_runOptions_enableTags :: [TableSpecTag]

_runOptions_disableTags :: [TableSpecTag]

parseRefTable :: ReadM (Ref Table)

parseLogLevel :: ReadM Severity

parseRunOptions :: Parser CliOptions

Parser for the run CLI sub-command

Parses available commands for the run sub-command and provides detail CLI help text

parseDumpOptions :: Parser CliOptions

Parser for the dump CLI sub-command

Parses available commands for the run sub-command and provides detail CLI help text

parseCliOption :: Parser CliOptions

parseSpecMetaArgs :: [ByteString] -> Either String SpecMetaArgs

dumpDeps :: ParseSQL bk => (AppConfig -> ResourceT IO ComboEnv) -> bk -> [(SpecRuntime bk, Spec bk a)] -> IO ()

compileSpecs :: MonadIO m => [(t, Spec b a)] -> m [(t, Specs b)]

Compile given list of specs into ready to execute form.

specMain :: (SpecBackend bk drop t1 t2 t3 t4 t5 t6 t7, RenderSql (CreateTableAs (BackendTableMeta bk)) bk, RenderSql CreateViewAs bk, ParseSQL bk, Backend bk) => bk -> [(SpecRuntime bk, Spec bk a)] -> IO ExitCode

Use this to make your main entry point for your spec running application. specMain :: bk -> [(NS.SpecRuntime bk, NST.Spec bk ())] -> IO ()

specMainWith

Arguments

:: forall bk drop t1 t2 t3 t4 t5 t6 t7 a. (SpecBackend bk drop t1 t2 t3 t4 t5 t6 t7, RenderSql (CreateTableAs (BackendTableMeta bk)) bk, RenderSql CreateViewAs bk, ParseSQL bk, Backend bk)

=> (AppConfig -> ResourceT IO ComboEnv)
How to build the combo. register any cleanup

-> CliOptions

-> bk

-> CustomSpecPreprocessor bk

-> CustomValidator

-> [(SpecRuntime bk, Spec bk a)]

-> IO ExitCode

A more customizable specMain for embedding in larger programs. This API is unstable

exitRunFailed :: ExitCode

exitValidationFailed :: ExitCode

allocComboEnv :: AppConfig -> ResourceT IO ComboEnv