Napkin.Spec.CliReporter

Documentation

data TableState

Constructors

Pending

Processing

Skipped

Done

Failed

data UIName

Unique names for user interface elements necessary for brick.

Constructors

SpecListViewport

SpecListView

data PreviousStates

Constructors

PreviousStates

Fields

_previousStates_processingTime :: Maybe ElapsedSeconds

_previousStates_finishedTime :: Maybe ElapsedSeconds

_previousStates_info :: Maybe String

freshPreviousStates :: PreviousStates

prevStatesFinishedTime :: PreviousStates -> Maybe ElapsedSeconds

type ElapsedSeconds = Integer

type TableStates = Map String (TableState, PreviousStates)

data DrawState

Constructors

DrawState

Fields

startTime :: UTCTime

tableStates :: TableStates

elapsedTime :: ElapsedSeconds

processingTime :: ElapsedSeconds

runDone :: Bool

wantsToQuit :: Bool

eta :: ElapsedSeconds

trackingQueryStats :: TrackingQueryStats

allTablesProcessed :: TableStates -> Bool

True iff all tables have been serviced so that the elapsed time can stop ticking.

type TableName = String

type ErrorInfo = String

type SuccessInfo = String

data UpdateEvent

Constructors

TableProcessingEvent TableName

TableDoneEvent TableName TrackingQueryStats

TableFailedEvent TableName ErrorInfo

TableSkippedEvent TableName

RunDone TrackingQueryStats

FinishFatal

Tick

eventHandler :: DrawState -> BrickEvent UIName UpdateEvent -> EventM UIName (Next DrawState)

cliReporter :: RuntimeEventQueue -> [String] -> ElapsedSeconds -> IO Success

Pump the runtime event queue to display Spec execution progress in the terminal. Returns after all Specs are completed.

success :: DrawState -> Success

prettyElapsedTime :: Integer -> String

printFinalReport :: DrawState -> IO ()

filterByState :: (TableState -> Bool) -> TableStates -> TableStates

drawScreen :: DrawState -> [Widget UIName]

prettyUTCTime :: UTCTime -> String

data TableStateLine

Constructors

TableStateLine String TableState PreviousStates

sortedStatesList :: TableStates -> [TableStateLine]

renderOneState :: ElapsedSeconds -> TableStateLine -> Widget UIName

cliAttrMap :: AttrMap

successStateLine :: AttrName

failureStateLine :: AttrName

processingStateLine :: AttrName

pendingStateLine :: AttrName

introBox :: AttrName

napkinText :: [String]