Napkin.Spec.ComboTypes

Documentation

data AppConfig

Constructors

AppConfig

Fields

_acAppName :: Text

_acLog :: Severity

_acSqlitePath :: Maybe FilePath

_acInteractive :: Bool

data RuntimeEventType

Maintains some basic, human-reportable state over the run of the Spec in order to show progress.

Constructors

RuntimeEventStart
generated when a Spec starts to run

RuntimeEventDone TrackingQueryStats
generated when a Spec successfully completes

RuntimeEventFailed String
generated when a Spec throws an error

RuntimeEventSkipped
generated when a Spec update is skipped

data RuntimeEvent

Constructors

RuntimeEvent (Ref Table) RuntimeEventType

RuntimeEventComplete TrackingQueryStats
generated when napkin is done running

RuntimeEventFatal
generated from the napkin backend for fatal events (such as failed validation)

type RuntimeEventQueue = TBQueue RuntimeEvent

data ComboEnv

Constructors

ComboEnv

Fields

_ceAppConfig :: AppConfig

_ceLog :: LogEnv

_ceGH :: Pool Sqlite

_ceEventQueue :: Maybe RuntimeEventQueue

newtype ComboT m a

Constructors

ComboT

Fields

getComboT :: RWST ComboEnv () TrackingQueryStats m a

type Combo = ComboT IO

Versions of the FromJSON helpers that use Typeable to cut down boilerplate

withObject' :: forall a. Typeable a => (Object -> Parser a) -> Value -> Parser a

Functions like withObject can be used to provide nice type error messages rather than a non-exhaustive pattern match on Value:

> instance FromJSON Foo where
>   parseJSON = withObject' parseFoo
>     where
>       parseFoo v = Foo <$> v .: "bar" <*> v .: "baz"

(.:?!) :: FromJSON a => Object -> Text -> Parser (Maybe a)

Restores previous behavior of .:? from aeson < 0.10: the field can either be null or completely omitted. This works by turning your parser into a Maybe a which is itself optional, thus (Parser (Maybe (Maybe a)) and then flattens the result into 1 level.

class HasAppConfig c where

Minimal complete definition

appConfig

Methods

appConfig :: Lens' c AppConfig

acAppName :: Lens' c Text

acInteractive :: Lens' c Bool

acLog :: Lens' c Severity

acSqlitePath :: Lens' c (Maybe FilePath)

ceLog :: Lens' ComboEnv LogEnv

ceGH :: Lens' ComboEnv (Pool Sqlite)

ceEventQueue :: Lens' ComboEnv (Maybe RuntimeEventQueue)

ceAppConfig :: Lens' ComboEnv AppConfig