Napkin.Spec.Dump

Documentation

data SqlFileType

Constructors

SqlFileType_Query

SqlFileType_InsertStatement

specToSQLDir :: (RenderSql (CreateTableAs d) b, RenderSql Query b, RenderSql CreateViewAs b, ParseSQL b, BackendTableMeta b ~ d, Default d) => b -> Specs b -> IncludeUnmanagedTables -> SqlFileType -> FilePath -> IO ()

Main entrypoint into SQL directory generation.

specGraphAsDot :: SpecDepGraph -> DotGraph Int

tableau10Colors :: [Color]

tableau20Colors :: [Color]

specGraphAsDotText :: SpecDepGraph -> Text

specGraphAsSQLDirectory :: (RenderSql (CreateTableAs d) b, RenderSql Query b, RenderSql CreateViewAs b, ParseSQL b, BackendTableMeta b ~ d, Default d) => b -> SpecDepGraph -> SpecTableMap b -> GenerateInserts -> FilePath -> IO ()

specGraphAsSQL :: (RenderSql (CreateTableAs d) b, RenderSql Query b, RenderSql CreateViewAs b, ParseSQL b, Default (BackendTableMeta b), BackendTableMeta b ~ d) => b -> SpecDepGraph -> SpecTableMap b -> GenerateInserts -> IO [(Ref Table, [String])]

Returns a list of SQL sources which can be used to construct the database in order of dependency.

type GenerateInserts = Bool

renderTableSpec :: forall b d. (RenderSql Query b, RenderSql CreateViewAs b, RenderSql (CreateTableAs d) b, ParseSQL b, Default (BackendTableMeta b), BackendTableMeta b ~ d) => b -> TableSpec b -> GenerateInserts -> IO [String]

Render a TableSpec using a BigQuery backend. Requires IO to perform Raw SQL file reading. GenerateInserts can indicate whether the generated SQL files should include INSERT statements. This is used for breaking up BigQuery statements which would otherwise be too long to execute.

Orphan instances

Labellable SpecDepGraphLabel

Instance details

Methods

toLabelValue :: SpecDepGraphLabel -> Label