Napkin.Spec.Persist

Documentation

loadSpecState :: PersistBackend m => Pipeline -> m SpecState

srLastRun :: Traversal' SpecRun UTCTime

srDepChain :: Traversal' SpecRun DepChain

type SpecRunId = Int

newSpecRunId :: PersistBackend m => m SpecRunId

recordedRun :: PersistBackend m => Pipeline -> SpecRunId -> Ref Table -> DepChain -> (forall b. m b -> ComboT IO b) -> ComboT IO a -> ComboT IO (Either SomeException a, SpecRun)

Do a run, recording the instance in the database. Return the recorded instance along with the run results.

expectedRuntimeForGraph :: PersistBackend m => SpecDepGraph -> m Int

longestPathLength :: Real b => Gr a b -> Node -> Node -> b

averageTotalRuntime :: PersistBackend m => m Int

collect

Arguments

:: (Ord k, Foldable t)

=> (a -> k)
key in collected map

-> (a -> v)
value in collected map

-> (v -> v -> v)
collapse function for value

-> t a
something I can traverse, e.g. [a]

-> Map k v