Napkin.Spec.Types

Documentation

type Success = Bool

newtype Pipeline

A namespace for a given pipeline

Constructors

Pipeline

Fields

_unNamespace :: Text

type DepChain = Map (Ref Table) UTCTime

newtype SpecState

Constructors

SpecState

Fields

_ssTables :: Map (Ref Table) SpecRun

data SpecRun

Each run of the spec for a table is captured

Constructors

SpecRun

Fields

srApp :: Pipeline

srRunId :: Maybe Int

srTable :: Ref Table

srStarted :: UTCTime

srComplete :: Maybe UTCTime

srAction :: Maybe Text

srError :: Maybe Text

srRowsAffected :: Maybe Int64

srBytesAffected :: Maybe Int64

srCost :: Maybe Int64

cost in U.S. pennies

srServerTime :: Maybe Int64

milliseconds of server time required which may be different from the client's time persepective

srHookRowsAffected :: Maybe Int64

rows affected by pre- and post-hooks for this spec

srHookBytesAffected :: Maybe Int64

srHookCost :: Maybe Int64

srHookServerTime :: Maybe Int64

srDeps :: Maybe (Showing DepChain)

srDuration :: SpecRun -> NominalDiffTime

newtype Showing a

Constructors

Showing

Fields

_unShowing :: a

data Hook b

Constructors

AssertEqualCount Relation Relation
Fail iff the tables' row counts differ.

AssertApproxEqualCount Double Relation Relation
Fail iff the tables' row counts differ by more than the percentage fudge factor.

AssertGreaterThanCount Relation Relation
Fail iff the first table's row count is less than the second table's row count.

AssertGreaterThanConstant Relation Integer
Fail iff the table's row count is less than or equal to the constant.

AssertUniqueBy Relation [(SExp, [Ref SExp])]
Fail if there are non-unique rows in the result for the multiple predicate and column names.

AssertSExp Relation SExp
Fail if is the SExp is false or NULL.

AssertNotNull Relation [SExp]
Fail if any of the given expressions/columns have more than 1 NULL row.

CustomHook QueryTransformer (BackendConn b -> QueryTransformer -> ComboT IO ())
Write your own Hook.

assertUniqueBy :: [Ref SExp] -> Relation -> Hook b

runProcessHook :: String -> [String] -> Hook b

data SpecTarget

Constructors

SpecTable

SpecView

type IsManagedTable = Bool

data QueryTransformer

Constructors

QueryTransformer

Fields

_qt_refs :: Ref Table -> Ref Table

renamer for table names for manual cases

_qt_query :: Query -> Query

a packaged up query transform that namespaces all names

_qt_func :: Ref Function -> Ref Function

qt_refs :: Lens' QueryTransformer (Ref Table -> Ref Table)

qt_query :: Lens' QueryTransformer (Query -> Query)

qt_func :: Lens' QueryTransformer (Ref Function -> Ref Function)

applyQueryTransformerQueries :: Data b => QueryTransformer -> b -> b

Apply renamer (e.g. for namespacing) for all Query objects inside

applyQueryTransformerFunctions :: Data b => QueryTransformer -> b -> b

applyQueryTransformerToRef :: TableRef b => QueryTransformer -> b -> b

applyQueryTransformerToQuery :: QueryTransformer -> Query -> Query

data SqlSource

Constructors

HaskellSqlSource

HaskellScriptSqlSource HaskellModuleSource HaskellFunctionSource

FileSqlSource FilePath

type HaskellModuleSource = Text

type HaskellFunctionSource = Text

data CreateAction b

Constructors

WithQuery
Due to backend query size limitations, additional queries can be split off and added here. The renamer is required because the Query maybe be a raw SQL string which requires post-processing.

Fields

_createAction_withQuery_q1 :: Query

_createAction_withQuery_qs :: [Query]

_createAction_withQuery_targetType :: SpecTarget

_createAction_withQuery_transformer :: QueryTransformer

_createAction_withQuery_source :: SqlSource

_createAction_withQuery_deps :: [Ref Table]

a query may have additional dependencies beyond the deps that can be inferred due to table wildcards

_createAction_withQuery_blacklistDeps :: [Ref Table]

a query may reference itself in a cycle or it wishes napkin not to wait for a table to be updated- the blacklist breaks the cycle to allow napkin to continue

CreateRawSql
Raw SQL file, deps, target and a content transformer prior to run

Fields

_createAction_createRawSql_file :: FilePath

_createAction_createRawSql_templateVars :: SqlTemplateVariables

_createAction_createRawSql_deps :: [Ref Table]

_createAction_createRawSql_blacklistDeps :: [Ref Table]

remove a dependency to explicitly allow for cycles

_createAction_createRawSql_targetType :: SpecTarget

_createAction_createRawSql_transformer :: QueryTransformer

CustomCreate
Custom action with explicit list of other table dependencies

Fields

_createAction_customCreate_deps :: [Ref Table]

_createAction_customCreate_action :: TableStatus -> QueryTransformer -> SpecMetaArgs -> BackendConn b -> ComboT IO ()

data TableSpec b

Constructors

TableSpec

Fields

_specTable :: Ref Table

_specTableOptions :: Maybe (BackendTableMeta b)

_specAction :: CreateAction b

_specPreHooks :: [Hook b]

_specPostHooks :: [Hook b]

_specUpdate :: [UpdateStrategy]

_specGrants :: [([Privilege], [Actor])]

_specTags :: [TableSpecTag]

allows subgraph execution

type TableSpecTag = String

data UpdateStrategy

Each strategy acts independently. For example, if you have only UpdateWithDependency, it wouldn't update even if the table were missing. You need to specify each strategy in the list.

Constructors

UpdatePeriodically NominalDiffTime

UpdateIfMissing

UpdateAlways

UpdateWithDependency

data TableStatus

Status when a table is about to be evaluated for update.

Constructors

TableStatus

Fields

_tstatusLastUpdate :: Maybe SpecRun

Last time this table was run

_tstatusExists :: Bool

_tstatusDeps :: DepChain

Current status of the dependencies

type SpecTableMap b = Map (Ref Table) (TableSpec b)

type SpecMetaArgs = Map Text Value

data Specs b

Constructors

Specs

Fields

_specsTables :: SpecTableMap b

_specsHooks :: [([Ref Table], Hook b)]

independent hooks that can run as soon as given dependencies are ready

_specsMetaArgs :: SpecMetaArgs

arguments which can be passed to custom creators

newtype Spec b a

Constructors

Spec

Fields

_runSpec :: StateT (Specs b) IO a

runSpec :: Spec b a -> IO (Specs b)

class HasSpec s bk where

Methods

pullSpec :: s -> IO (Specs bk)

ssTables :: Iso' SpecState (Map (Ref Table) SpecRun)

unShowing :: forall a a. Iso (Showing a) (Showing a) a a

srTable_ :: Lens' SpecRun (Ref Table)

srStarted_ :: Lens' SpecRun UTCTime

srServerTime_ :: Lens' SpecRun (Maybe Int64)

srRunId_ :: Lens' SpecRun (Maybe Int)

srRowsAffected_ :: Lens' SpecRun (Maybe Int64)

srHookServerTime_ :: Lens' SpecRun (Maybe Int64)

srHookRowsAffected_ :: Lens' SpecRun (Maybe Int64)

srHookCost_ :: Lens' SpecRun (Maybe Int64)

srHookBytesAffected_ :: Lens' SpecRun (Maybe Int64)

srError_ :: Lens' SpecRun (Maybe Text)

srDeps_ :: Lens' SpecRun (Maybe (Showing DepChain))

srCost_ :: Lens' SpecRun (Maybe Int64)

srComplete_ :: Lens' SpecRun (Maybe UTCTime)

srBytesAffected_ :: Lens' SpecRun (Maybe Int64)

srApp_ :: Lens' SpecRun Pipeline

srAction_ :: Lens' SpecRun (Maybe Text)

specUpdate :: forall b. Lens' (TableSpec b) [UpdateStrategy]

specTags :: forall b. Lens' (TableSpec b) [TableSpecTag]

specTableOptions :: forall b. Lens' (TableSpec b) (Maybe (BackendTableMeta b))

specTable :: forall b. Lens' (TableSpec b) (Ref Table)

specPreHooks :: forall b. Lens' (TableSpec b) [Hook b]

specPostHooks :: forall b. Lens' (TableSpec b) [Hook b]

specGrants :: forall b. Lens' (TableSpec b) [([Privilege], [Actor])]

specAction :: forall b. Lens' (TableSpec b) (CreateAction b)

tstatusLastUpdate :: Lens' TableStatus (Maybe SpecRun)

tstatusExists :: Lens' TableStatus Bool

tstatusDeps :: Lens' TableStatus DepChain

specsTables :: forall b. Lens' (Specs b) (SpecTableMap b)

specsMetaArgs :: forall b. Lens' (Specs b) SpecMetaArgs

specsHooks :: forall b. Lens' (Specs b) [([Ref Table], Hook b)]

createAction_withQuery_transformer :: forall b. Traversal' (CreateAction b) QueryTransformer

createAction_withQuery_targetType :: forall b. Traversal' (CreateAction b) SpecTarget

createAction_withQuery_source :: forall b. Traversal' (CreateAction b) SqlSource

createAction_withQuery_qs :: forall b. Traversal' (CreateAction b) [Query]

createAction_withQuery_q1 :: forall b. Traversal' (CreateAction b) Query

createAction_withQuery_deps :: forall b. Traversal' (CreateAction b) [Ref Table]

createAction_withQuery_blacklistDeps :: forall b. Traversal' (CreateAction b) [Ref Table]

createAction_customCreate_deps :: forall b. Traversal' (CreateAction b) [Ref Table]

createAction_customCreate_action :: forall b b. Traversal (CreateAction b) (CreateAction b) (TableStatus -> QueryTransformer -> SpecMetaArgs -> BackendConn b -> ComboT IO ()) (TableStatus -> QueryTransformer -> SpecMetaArgs -> BackendConn b -> ComboT IO ())

createAction_createRawSql_transformer :: forall b. Traversal' (CreateAction b) QueryTransformer

createAction_createRawSql_templateVars :: forall b. Traversal' (CreateAction b) SqlTemplateVariables

createAction_createRawSql_targetType :: forall b. Traversal' (CreateAction b) SpecTarget

createAction_createRawSql_file :: forall b. Traversal' (CreateAction b) FilePath

createAction_createRawSql_deps :: forall b. Traversal' (CreateAction b) [Ref Table]

createAction_createRawSql_blacklistDeps :: forall b. Traversal' (CreateAction b) [Ref Table]

allSpecsTables :: Specs b -> [Ref Table]

namespaceTables :: forall b. (Ref Table -> Ref Table) -> Specs b -> Specs b

Rename all napkin-managed tables in the `Specs b`.

namespaceAllTables :: forall b. (Ref Table -> Ref Table) -> Specs b -> Specs b

Apply a renamer to all tables in the `Specs b` (managed or unmanaged).

namespaceFunctions :: forall b. (Ref Function -> Ref Function) -> Specs b -> Specs b

setAllTableGrants :: MonadState (Specs b) m => [([Privilege], [Actor])] -> m ()

Set the same Grant permissions for all tables defined in the spec at once. Convenience function for a common use case. Acts monadically, so it will apply to all specs so far defined in the Spec block.

data SpecRunConstructor v

Constructors

v ~ ConstructorMarker SpecRun => SpecRunConstructor

migrateDb :: PersistBackend m => m ()

safeCopy :: SafeCopy a => Prism' ByteString a

zlib compresses sagecopy blob

specDeps :: TableSpec b -> [Ref Table]

preprocessDependencies :: (MonadIO m, ParseSQL b) => b -> Specs b -> m (Specs b)

add dependencies to CreateRawSql actions

prettyQueryStats :: QueryStats -> String

Orphan instances

ToJSON (Ref a)

Instance details

Methods

toJSON :: Ref a -> Value

toEncoding :: Ref a -> Encoding

toJSONList :: [Ref a] -> Value

toEncodingList :: [Ref a] -> Encoding

NeverNull (Ref a)

Instance details

PersistField (Ref a)

Instance details

Methods

persistName :: Ref a -> String

toPersistValues :: PersistBackend m => Ref a -> m ([PersistValue] -> [PersistValue])

fromPersistValues :: PersistBackend m => [PersistValue] -> m (Ref a, [PersistValue])

dbType :: DbDescriptor db => proxy db -> Ref a -> DbType

SinglePersistField (Ref a)

Instance details

Methods

toSinglePersistValue :: PersistBackend m => Ref a -> m PersistValue

fromSinglePersistValue :: PersistBackend m => PersistValue -> m (Ref a)

PurePersistField (Ref a)

Instance details

Methods

toPurePersistValues :: Ref a -> [PersistValue] -> [PersistValue]

fromPurePersistValues :: [PersistValue] -> (Ref a, [PersistValue])

PrimitivePersistField (Ref a)
Instances

Instance details

Methods

toPrimitivePersistValue :: Ref a -> PersistValue

fromPrimitivePersistValue :: PersistValue -> Ref a