Napkin.Spec.Yaml.Interpret

Description

This module should be seen as an interpreter that takes the YAML version of a spec and converts it to a proper, fully typed Spec value.

In a sense, it's also a bit of a typechecker: YAML versions are necessarily dynamically typed and there's the possibility of type mismatches.

Documentation

data YamlInterpretError

Constructors

YamlInterpretError_MismatchingBacked YamlBackend YamlBackend

YamlInterpretError_MismatchingTableMeta YamlTableMeta

YamlInterpretError_ParseError ParseException

loadYaml :: ByteString -> Either YamlInterpretError YamlSpec

interpretBackend :: forall b m. (ReifiesBackend b, ToYamlBackend b, Monad m) => YamlBackend -> ExceptT YamlInterpretError m ()

Validate that caller's desired backend matches against what was specified in the YAML file.

interpretYamlToSpec :: forall bk t1 t2 t3 t4. (ReifiesBackend bk, ToYamlBackend bk, Typeable bk, Command (CreateTableAs (BackendTableMeta bk)) t1 bk, Command CheckTableExists Bool bk, Command InsertIntoQuery t1 bk, Command DeleteFrom t2 bk, Command DropTable t3 bk, Command RenameTable t4 bk, Drop (CreateTableAs (BackendTableMeta bk)) DropTable, RenderSql SExp bk, ParseSQL bk, Default (BackendTableMeta bk), Typeable (BackendTableMeta bk)) => FilePath -> YamlSpec -> bk -> [FilePath] -> ExceptT YamlInterpretError (Spec bk) ()

The main entry point interpreting a given YAML spec file as a Spec value ready to run via executeSpec

This function is intended to unroll/interpet as much as possible to get the spec in its final state ready to be actually run.

interpretTable :: forall bk t1 t2 t3 t4. (Typeable (BackendTableMeta bk), Command (CreateTableAs (BackendTableMeta bk)) t1 bk, Command CheckTableExists Bool bk, Command InsertIntoQuery t1 bk, Command DeleteFrom t2 bk, Command DropTable t3 bk, Command RenameTable t4 bk, Drop (CreateTableAs (BackendTableMeta bk)) DropTable, RenderSql SExp bk, Default (BackendTableMeta bk), ParseSQL bk, ReifiesBackend bk) => YamlBackend -> FilePath -> FilePath -> [String] -> [FilePath] -> YamlTableSpec -> ExceptT YamlInterpretError (Spec bk) ()

Main entry point for fully interpreting a given table's entry in the YAML file.

interpretTableMeta :: forall bk m. (Monad m, Typeable (BackendTableMeta bk)) => YamlTableMeta -> ExceptT YamlInterpretError m (Maybe (BackendTableMeta bk))

Parse backend-specific table meta content from YAML file into the type dictated by the particular backend.

interpretHook :: YamlHook -> Hook bk

interpretAction :: forall bk m t1 t2 t3 t4. (Monad m, MonadIO m, Typeable (BackendTableMeta bk), Command (CreateTableAs (BackendTableMeta bk)) t1 bk, Command CheckTableExists Bool bk, Command InsertIntoQuery t1 bk, Command DeleteFrom t2 bk, Command DropTable t3 bk, Command RenameTable t4 bk, Drop (CreateTableAs (BackendTableMeta bk)) DropTable, RenderSql SExp bk, Default (BackendTableMeta bk), ParseSQL bk, ReifiesBackend bk) => YamlBackend -> Maybe (BackendTableMeta bk) -> Ref Table -> FilePath -> FilePath -> [String] -> [FilePath] -> YamlCreateAction -> ExceptT YamlInterpretError m (CreateAction bk)

yamlSpecTarget :: YamlSpecTarget -> SpecTarget

interpretSpec :: forall bk t1 t2 t3 t4. (Typeable (BackendTableMeta bk), Typeable bk, Command (CreateTableAs (BackendTableMeta bk)) t1 bk, Command CheckTableExists Bool bk, Command InsertIntoQuery t1 bk, Command DeleteFrom t2 bk, Command DropTable t3 bk, Command RenameTable t4 bk, Drop (CreateTableAs (BackendTableMeta bk)) DropTable, RenderSql SExp bk, Default (BackendTableMeta bk), ParseSQL bk, ReifiesBackend bk) => YamlBackend -> FilePath -> FilePath -> [String] -> [FilePath] -> YamlHaskellSymbol -> ExceptT YamlInterpretError (Spec bk) ()