Napkin.Spec.Yaml.Types

Documentation

data YamlSpec

Constructors

YamlSpec

Fields

_yamlSpec_sqlFolder :: FilePath

indicate folders within project to scan for SQL and other source files. By default should be sql

_yamlSpec_connectionURL :: String

URL to indicate how to create the database connection

_yamlSpec_haskellFolder :: Maybe FilePath

folder used to find the Haskell source code. When not provided, assumed to be haskell/

_yamlSpec_tableSpecs :: [YamlTableSpec]

_yamlSpec_backend :: YamlBackend

indicate backend to be used in this spec

_yamlSpec_specPreprocessor :: Maybe YamlHaskellSymbol

specify a function to run as the napkin spec preprocessor

_yamlSpec_validator :: Maybe YamlHaskellSymbol

specify a function to run as the napkin validator

_yamlSpec_extraHaskellPackages :: [String]

specify additional packages which should be exposed to the haskell functions in the YAML spec

_yamlSpec_haskellSpec :: Maybe YamlHaskellSymbol

specify an additional Haskell function to run in the Spec monad after all the TableSpecs.

data YamlBackend

We include a set of default choices user can supply, which helps us disambiguate database, dataset/schema and similar choices when left unspecified in SQL files.

Constructors

YamlBackend_BigQuery (Maybe YamlBigQueryDefaults)

YamlBackend_Postgres (Maybe YamlPostgresDefaults)

YamlBackend_Redshift (Maybe YamlPostgresDefaults)

data YamlBigQueryDefaults

A set of defaults that's useful to capture from the user to disambiguate ad-hoc queries.

Constructors

YamlBigQueryDefaults

Fields

_yamlBigQueryDefaults_project :: Text

_yamlBigQueryDefaults_dataset :: Text

data YamlPostgresDefaults

A set of defaults that's useful to capture from the user - allows us to run SQL ambiguous in database/schema, or run without asking an explicit target server from the user.

We may also want to bake in optional defaults on username, but for now I'm keeping that separate to be pulled in from authentication info meant to reside separately from the spec config.

For now, we're keeping all Postgres based cases unified. We may break them out as needed in the future.

Constructors

YamlPostgresDefaults

Fields

_yamlPostgresDefaults_server :: Text

_yamlPostgresDefaults_database :: Text

_yamlPostgresDefaults_schema :: Text

class ToYamlBackend b where

Methods

toYamlBackend :: YamlBackend

data YamlTableSpec

Constructors

YamlTableSpec

Fields

_yamlTableSpec_targetTable :: Ref Table

_yamlTableSpec_targetTableOptions :: YamlTableMeta

_yamlTableSpec_createAction :: YamlCreateAction

_yamlTableSpec_preHooks :: [YamlHook]

_yamlTableSpec_postHooks :: [YamlHook]

_yamlTableSpec_updateStrategy :: [UpdateStrategy]

_yamlTableSpec_tags :: [TableSpecTag]

data YamlTableMeta

Table options specific to each backend.

Constructors

YamlTableMeta_BigQuery (BackendTableMeta BigQuery)

YamlTableMeta_Redshift (BackendTableMeta Redshift)

YamlTableMeta_None

data YamlCreateAction

Constructors

YamlCreateAction_Mustache
Create a table spec from an sql file

Fields

_yamlCreateAction_mustache_file :: FilePath

_yamlCreateAction_mustache_variables :: SqlTemplateVariables

_yamlCreateAction_mustache_deps :: [Ref Table]

_yamlCreateAction_mustache_hiddenDeps :: [Ref Table]

_yamlCreateAction_mustache_targetType :: YamlSpecTarget

_yamlCreateAction_mustache_strategy :: YamlCreateStrategy

YamlCreateAction_Haskell
Create a table spec from a Haskell function in the haskell folder

Fields

_yamlCreateAction_haskell_symbol :: YamlHaskellSymbol

_yamlCreateAction_haskell_targetType :: YamlSpecTarget

data YamlHaskellSymbol

Constructors

YamlHaskellSymbol Text Text

data YamlHook

TODO: Flesh these out for more typical use cases.

Using generalized napkin types here; maybe we can let the user write little snippets into YAML for increased power.

Constructors

YamlHook_AssertEqualCount Relation Relation

YamlHook_AssertUniqueBy Relation [Ref SExp]

YamlHook_AssertNotNull Relation [SExp]

data YamlSpecTarget

Constructors

YamlSpecTarget_Table

YamlSpecTarget_View

data YamlCreateStrategy

allows selection of wrappers such as incrementalByTimeRawSql

Constructors

YamlCreateStrategy_IncrementalByTime (Ref SExp) Day Integer

YamlCreateStrategy_IncrementalByPK YamlHaskellSymbol (NonEmpty (Ref SExp)) Day Integer

YamlCreateStrategy_Default

Orphan instances

ToJSON UpdateStrategy

Instance details

Methods

toJSON :: UpdateStrategy -> Value

toEncoding :: UpdateStrategy -> Encoding

toJSONList :: [UpdateStrategy] -> Value

toEncodingList :: [UpdateStrategy] -> Encoding

FromJSON SqlTemplateVariables

Instance details

Methods

parseJSON :: Value -> Parser SqlTemplateVariables

parseJSONList :: Value -> Parser [SqlTemplateVariables]

FromJSON UpdateStrategy

Instance details

Methods

parseJSON :: Value -> Parser UpdateStrategy

parseJSONList :: Value -> Parser [UpdateStrategy]

FromJSON (Ref a)

Instance details

Methods

parseJSON :: Value -> Parser (Ref a)

parseJSONList :: Value -> Parser [Ref a]