Napkin.Types

Documentation

data Name

Constructors

Specific String

Star

nameString :: Name -> String

_name :: Iso' String Name

name :: String -> Name

newtype Ref a

A way to refer to entities like fields, tables, schemas, etc. Namespacing is common so we keep a list here.

Constructors

Ref

Fields

_unRef :: NonEmpty Name

class IsRef a where

Class of string-like types that can be interpreted as Refs

Methods

ref :: forall b. a -> Ref b

refComponentCount :: Ref a -> Int

Return the number of Names in the Ref.

stringRefIso :: Iso' String (Ref a)

stringRef :: String -> Ref a

nameRef :: Name -> Ref a

namesRef :: [Name] -> Ref a

refString :: Ref t -> String

retagRef :: Ref t -> Ref a

appendRef :: Ref a -> Name -> Ref a

star :: SExp

data Table

data Function

data Selected a

Select something and give it a name; e.g. (my_field as foo)

Constructors

Selected

Fields

_selectRef :: Ref a

_selectItem :: a

data Alias a

Like selected, but aliasing is optional

Constructors

Alias

Fields

_aliasRef :: Maybe (Ref a)

_aliasItem :: a

nonAlias :: a -> Alias a

class WithName a where

Methods

as :: b -> Ref b -> a b

e.g. var "foo" as "bar"

sa :: WithName a => Ref b -> b -> a b

Flipped as

selectToAlias :: Selected a -> Alias a

data Value

Constructors

VDouble Double

VInt Int64

VBool Bool

VDate Day

VDateTime UTCTime

VText Text

VBinary ByteString

VInterval Interval

VNull

VDatePart DatePart

VArray [SExp]

newtype Interval

Intervals could internally make use of expressions

Constructors

Interval

Fields

_unInterval :: [(SExp, DatePart)]

data IntInterval

A simpler sub-variant to Interval that fits certain cases better. Maps onto Interval underneath.

Constructors

IntInterval SExp DatePart

data SExp

Constructors

Lit Value

Var (Ref SExp)

Extern ExternFun
external function call

Case [(SExp, SExp)] SExp
case with a final default

ExceptColumns SExp [Ref SExp]
t.* EXCEPT(a,b,c)

SubQuery Query

var :: Ref a -> SExp

varString :: String -> SExp

nil :: SExp

data NullStrategy

Constructors

IgnoreNulls

RespectNulls

data Distinctness

Constructors

Distinct

NonDistinct

type Partition = [SExp]

data WOver

Constructors

WOver

Fields

_overPartition :: Partition

_overOrder :: Order

_overFrame :: Maybe WindowFrame

data WindowFun

Constructors

SimpleWindow (Ref Function) [SExp] WOver (Maybe NullStrategy)

data WindowFrame

Constructors

WindowBetween FrameLength FrameLength

WindowFrom FrameLength

data FrameLength

Constructors

CurrentRow

Unbounded

LimitedPreceding Int

LimitedFollowing Int

data FunModifier

Constructors

ModDistinct Distinctness

ModNulls NullStrategy

ModOrder Order

ModLimit Int

ModOver WOver

data ExternFun

Constructors

SimpleExtern (Ref Function) [SExp]

ModExtern (Ref Function) [SExp] [FunModifier]

Cast SExp Type

SafeCast SExp Type

ExternWindow WindowFun

ExternRaw String

data Type

Constructors

TySmallInt

TyInteger

TyBigInt

TyDecimal

TyReal

TyDouble

TyBool

TyChar

TyVarChar

TyVarCharWithLen Int

TyDate

TyTimestamp

TyDatetime

data DatePart

Constructors

Millennium

Century

Decade

Epoch

Year

Quarter

Month

Week

DayOfWeek

DayOfYear

Day

Hour

Minute

Second

Millisecond

Microsecond

type Order = [OrderPart]

type GroupBy = [SExp]

data OrderDir

Constructors

Asc

Desc

invertOrder :: OrderDir -> OrderDir

Flip order

data NullOrder

Constructors

NullsFirst

NullsLast

data OrderPart

Constructors

OrderPart

Fields

_opExp :: SExp

_opOrder :: OrderDir

_opNulls :: Maybe NullOrder

order :: SExp -> OrderPart

data Relation

Constructors

TableRelation (Ref Table)

QueryRelation Query

ArrayRelation SExp

RawRelation [Ref Table] String
a raw relation with its dependencies

table :: Ref a -> Relation

treat a reference as relation

data JoinType

Constructors

JoinLeft

JoinRight

JoinInner

JoinCross

JoinOuter

data From

Constructors

From (Selected Relation) [(Selected Relation, JoinType, Either (Maybe SExp) [Ref SExp])]

fromAliases :: From -> [Ref Relation]

true :: SExp

false :: SExp

data UnionType

Constructors

SetUnion

DuplicateUnion

IntersectUnion

MinusUnion

data RawQuery

Constructors

RawQuery

Fields

_rawQuery_deps :: [Ref Table]

_rawQuery_raw :: String

type WithClauses = OMap (Ref Query) (Selected Query)

orderWithInserted :: WithClauses -> [(Ref Query, Selected Query)]

data Query

Constructors

Query

Fields

_queryWith :: WithClauses

_querySelect :: [Alias SExp]

_queryFrom :: Maybe From

_queryWhere :: Maybe SExp

_queryHaving :: Maybe SExp

_queryGroup :: GroupBy

_queryOrder :: Order

_queryLimit :: Maybe Int

_queryOffset :: Maybe Int

_queryDistinct :: Distinctness

Union

Fields

_unionType :: UnionType

_unionQuery1 :: Query

_unionQuery2 :: Query

_unionWiths :: WithClauses

unions may share with clauses

QueryRaw RawQuery
A raw query with all the dependencies contained therein

emptyQuery :: Query

mkUnion :: UnionType -> [Query] -> Query

Convenience: Combine list of queries via union of the same type

collectUnionQueries :: Query -> [Query]

Pull apart any UNIONs and extract a flat list of non-union queries.

data DefinedRelation

OO-style wrapper for objects that contain a full query definition

Constructors

forall s.(AsRelation s, HasDefinition s) => DefinedRelation s

data DefinedQuery

Like DefinedRelation, but only requires a defined query.

Constructors

forall s.HasDefinition s => DefinedQuery s

class AsRelation s where

Methods

asRelation :: s -> Relation

class HasDefinition s where

Methods

defQuery :: Lens s s Query Query

class MaybeQuery s where

Things that may or may not contain an explicit Query inside

Methods

getQuery :: s -> Maybe Query

queryFullySpecified :: Data a => a -> Bool

Has select * been used anywhere inside this thing? If so, False.

opOrder :: Lens' OrderPart OrderDir

opNulls :: Lens' OrderPart (Maybe NullOrder)

opExp :: Lens' OrderPart SExp

rawQuery_raw :: Lens' RawQuery String

rawQuery_deps :: Lens' RawQuery [Ref Table]

unionWiths :: Traversal' Query WithClauses

unionType :: Traversal' Query UnionType

unionQuery2 :: Traversal' Query Query

unionQuery1 :: Traversal' Query Query

queryWith :: Traversal' Query WithClauses

queryWhere :: Traversal' Query (Maybe SExp)

querySelect :: Traversal' Query [Alias SExp]

queryOrder :: Traversal' Query Order

queryOffset :: Traversal' Query (Maybe Int)

queryLimit :: Traversal' Query (Maybe Int)

queryHaving :: Traversal' Query (Maybe SExp)

queryGroup :: Traversal' Query GroupBy

queryFrom :: Traversal' Query (Maybe From)

queryDistinct :: Traversal' Query Distinctness

_QueryRaw :: Prism' Query RawQuery

_Union :: Prism' Query (UnionType, Query, Query, WithClauses)

_Query :: Prism' Query (WithClauses, [Alias SExp], Maybe From, Maybe SExp, Maybe SExp, GroupBy, Order, Maybe Int, Maybe Int, Distinctness)

_From :: Iso' From (Selected Relation, [(Selected Relation, JoinType, Either (Maybe SExp) [Ref SExp])])

_Star :: Prism' Name ()

_Specific :: Prism' Name String

_JoinOuter :: Prism' JoinType ()

_JoinCross :: Prism' JoinType ()

_JoinInner :: Prism' JoinType ()

_JoinRight :: Prism' JoinType ()

_JoinLeft :: Prism' JoinType ()

_RawRelation :: Prism' Relation ([Ref Table], String)

_ArrayRelation :: Prism' Relation SExp

_QueryRelation :: Prism' Relation Query

_TableRelation :: Prism' Relation (Ref Table)

unRef :: forall a a. Iso (Ref a) (Ref a) (NonEmpty Name) (NonEmpty Name)

_VArray :: Prism' Value [SExp]

_VDatePart :: Prism' Value DatePart

_VNull :: Prism' Value ()

_VInterval :: Prism' Value Interval

_VBinary :: Prism' Value ByteString

_VText :: Prism' Value Text

_VDateTime :: Prism' Value UTCTime

_VDate :: Prism' Value Day

_VBool :: Prism' Value Bool

_VInt :: Prism' Value Int64

_VDouble :: Prism' Value Double

_Microsecond :: Prism' DatePart ()

_Millisecond :: Prism' DatePart ()

_Second :: Prism' DatePart ()

_Minute :: Prism' DatePart ()

_Hour :: Prism' DatePart ()

_Day :: Prism' DatePart ()

_DayOfYear :: Prism' DatePart ()

_DayOfWeek :: Prism' DatePart ()

_Week :: Prism' DatePart ()

_Month :: Prism' DatePart ()

_Quarter :: Prism' DatePart ()

_Year :: Prism' DatePart ()

_Epoch :: Prism' DatePart ()

_Decade :: Prism' DatePart ()

_Century :: Prism' DatePart ()

_Millennium :: Prism' DatePart ()

_SubQuery :: Prism' SExp Query

_ExceptColumns :: Prism' SExp (SExp, [Ref SExp])

_Case :: Prism' SExp ([(SExp, SExp)], SExp)

_Extern :: Prism' SExp ExternFun

_Var :: Prism' SExp (Ref SExp)

_Lit :: Prism' SExp Value

selectRef :: forall a. Lens' (Selected a) (Ref a)

selectItem :: forall a. Lens' (Selected a) a

aliasRef :: forall a. Lens' (Alias a) (Maybe (Ref a))

aliasItem :: forall a. Lens' (Alias a) a

overPartition :: Lens' WOver Partition

overOrder :: Lens' WOver Order

overFrame :: Lens' WOver (Maybe WindowFrame)

class Val a where

Methods

val :: Prism' Value a

integral :: (Integral a, Integral b) => Iso' a b

_lit :: Val a => Prism' SExp a

lit :: Val a => a -> SExp

interval :: [(Double, DatePart)] -> SExp

Shorthand for using literal Doubles in making interval expressions.

txt :: String -> SExp

Monomorphic converter to make it easy to type string literals under OverloadedStrings

fullWindow :: WOver

Convenient starting point for full window (unbounded) for cases where not specifying window defaults to a partial window. (E.g. BigQuery)

class TableRef a where

Methods

tableRef :: Lens' a (Ref Table)

relationRef :: TableRef a => Getter a Relation

class HasDeps a where

Class of things that have external table dependencies

Methods

traverseDependencies :: Traversal' a (Ref Table)

tableDependencies :: HasDeps s => s -> [Ref Table]

queryDependencies :: Data s => Traversal' s (Ref Table)

Figure out external table dependencies within a construct - e.g. Query or one of the commands.

TODO Potential (unsure) bugs in here due to inner scopes spilling and clashing with names in outer scopes. Seems to work but can use proper extensive testing.

rawRelation :: HasDeps s => (s -> String) -> s -> Relation

Extract deps from object, render to string and wrap into a RawRelation

varAs :: Ref a -> Selected SExp

Polymorphic to support refs that come out of Relations, etc.

asSelf :: SExp -> Selected SExp

funAs :: (SExp -> b) -> Ref b -> Selected b

Apply function to a Ref and select it by the same name as the ref. Common use case in SELECT queries.

selectToRef :: Selected SExp -> SExp

Use the name of a Selected as a reference. Typically when you've computed a field in a subquery or a previous table, and you're now using that computation directly via its name.

selectedExps :: Traversal' [Selected a] a

selectedNames :: Traversal' [Selected a0] String

refName :: Applicative f => (String -> f String) -> Ref a -> f (Ref a1)

Pull the last name out of a Ref. E.g. if Ref contains a schema.table_name, pull just the table_name out.

refRoot :: Lens' (Ref a) Name

Get final segments on the ref path and grab the name

refJustRoot :: Ref a -> Ref a

refParentRoot :: Ref a -> ([String], Maybe String)

ne :: Iso' [a] (NonEmpty a)

selectedName :: Traversal' (Selected a) Name

atAlias :: (Applicative f, Choice p) => Ref t -> Optic' p f (Alias t) (Alias t)

aliasedNames :: HasDefinition s => Traversal' s (Ref SExp)

scopeRefs :: Data b => Ref t -> b -> b

Attach all column references within expression to given (table) reference.

modifyFromTables

Arguments

:: Data b

=> (Ref Table -> Ref Table)
A table name renaming function

-> b

-> b

Deep traverse structure and apply given name-changing function to each table's name but only if the table has been used in a FROM statement as a source table.

modifyFunctions :: Data b => (Ref Function -> Ref Function) -> b -> b

(^^.) :: Ref Relation -> Name -> SExp infixl 9

Attach given name on the relation

data AggLevel

Constructors

Unit
Expression is at the row level

Agg
Aggregate functions, like sum

Analytic
Analytic level (some backends)

data StatsType

Constructors

HookStats

ActionStats

data TrackingQueryStats

Used for tracking per-table and hook query statistics.

Constructors

TrackingQueryStats

Fields

_trackingQueryStats_tables :: Map (Ref Table, StatsType) QueryStats

tracks individual QueryStats for each table and per-table hooks

_trackingQueryStats_hooks :: QueryStats

tracks QueryStats for Spec-level hooks

_trackingQueryStats_unassociated :: QueryStats

tracks QueryStats which are not associated with a specific table

queryStatsForTable :: Ref Table -> TrackingQueryStats -> (Maybe QueryStats, Maybe QueryStats)

returns QueryStats, for the table creation spec and also for the spec's tables

totalQueryStatsForTable :: Ref Table -> TrackingQueryStats -> QueryStats

totalQueryStats :: TrackingQueryStats -> QueryStats

data QueryStats

Used for aggregate statistics gathered from running queries.

Constructors

QueryStats

Fields

_queryStats_rowsAffected :: !(Maybe Int64)

_queryStats_bytesAffected :: !(Maybe Int64)

_queryStats_cost :: !(Maybe Int64)

in U.S. pennies

_queryStats_time :: !(Maybe Int64)

milliseconds

Orphan instances

(Lift k, Lift v) => Lift (OMap k v :: Type)

Instance details

Methods

lift :: OMap k v -> Q Exp

liftTyped :: OMap k v -> Q (TExp (OMap k v))

Ord k => Semigroup (OMap k v)

Instance details

Methods

(<>) :: OMap k v -> OMap k v -> OMap k v

sconcat :: NonEmpty (OMap k v) -> OMap k v

stimes :: Integral b => b -> OMap k v -> OMap k v

Ord k => Monoid (OMap k v)

Instance details

Methods

mempty :: OMap k v

mappend :: OMap k v -> OMap k v -> OMap k v

mconcat :: [OMap k v] -> OMap k v

(NFData k, NFData v) => NFData (OMap k v)

Instance details

Methods

rnf :: OMap k v -> ()

Ord k => Ixed (OMap k a)

Instance details

Methods

ix :: Index (OMap k a) -> Traversal' (OMap k a) (IxValue (OMap k a))

Ord k => At (OMap k a)

Instance details

Methods

at :: Index (OMap k a) -> Lens' (OMap k a) (Maybe (IxValue (OMap k a)))