Napkin.Types.BigQuery

Documentation

data BigQuery

Constructors

BigQuery

newtype BQProjectId

Constructors

BQProjectId

Fields

unBQProjectId :: Text

newtype BQDatasetId

Constructors

BQDatasetId

Fields

unBQDatasetId :: Text

data BigQueryType

Represents the non-recursive bigquery types https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types

Constructors

BQString

BQBytes

BQDate

BQDateTime

BQTime

BQInt64

BQFloat64

BQBool

BQTimestamp

bqTypeString :: (IsString s, Eq s) => Prism' s BigQueryType

data GCSBucket

https://cloud.google.com/storage/docs/naming Between 3 and 63 chars, all lowercase, beginning and ending with alphanum, middle can contain alphanum or -,'_', and .. Avoid . in the bucket name as it requires verification. Also, common misspellings of google and IP addreses are not allowed but the risk of that seemed low so I omitted those rules.

gcsBucket :: GCSBucket -> Text

mkGCSBucket :: Text -> Either GCSBucketNameError GCSBucket

data GCSBucketNameError

Omits rules about IP address bucket names or close misspellings of google.

Constructors

GCSBBadChars !(Set Char)

GCSBInvalidStart !Char

GCSBInvalidEnd !Char

GCSBTooShort

GCSBTooLong

data GCSKey

https://cloud.google.com/storage/docs/naming No newlines, between 1 and 1024 chars

gcsKey :: GCSKey -> Text

mkGCSKey :: Text -> Either GCSKeyError GCSKey

data GCSKeyError

Constructors

GCSKTooShort

GCSKTooLong

GCSKBadChars !(Set Char)

newtype BQTableId

Constructors

BQTableId

Fields

unBQTableId :: Text

data GCSACLEntity

An entity which holds an ACL for an object or bucket in google cloud storage. I believe this means that whoever the entity targets can only hold one ACL, so you use this to do targeted/idempotent insertsreplacements of permissions on objectsbuckets.

Constructors

ACLEUserId !GCUserId

ACLEUserEmail !GCUserEmail

ACLEGroupId !GCGroupId

ACLEGroupEmail !GCGroupEmail

ACLEDomain !GCDomain

ACLEProject !GCTeam !GCProjectId

ACLEAllUsers

ACLEAllAuthenticatedUsers

gcsACLEntityText :: Prism' Text GCSACLEntity

data GCSACLRole

Constructors

ACLReader

ACLWriter

ACLOwner

gcsACLRoleText :: Prism' Text GCSACLRole

newtype GCUserId

Constructors

GCUserId

Fields

gcUserID :: Text

newtype GCUserEmail

Constructors

GCUserEmail

Fields

gcUserEmail :: Text

newtype GCGroupId

Constructors

GCGroupId

Fields

gcGroupId :: Text

newtype GCGroupEmail

Constructors

GCGroupEmail

Fields

gcGroupEmail :: Text

newtype GCDomain

Constructors

GCDomain

Fields

gcDomain :: Text

newtype GCTeam

Constructors

GCTeam

Fields

gcTeam :: Text

newtype GCProjectId

Constructors

GCProjectId

Fields

gcProjectId :: Text

data TableMeta

Constructors

TableMeta

Fields

_tableMeta_partitionCluster :: Maybe (ColumnName, [ColumnName])

_tableMeta_writeDisposition :: Maybe WriteDisposition

type ColumnName = String

data WriteDisposition

Constructors

WriteDisposition_WriteEmpty

WriteDisposition_WriteAppend

WriteDisposition_WriteTruncate

emptyTableMeta :: TableMeta

tableMeta :: ColumnName -> [ColumnName] -> TableMeta

tableMeta_partitionCluster :: Lens' TableMeta (Maybe (ColumnName, [ColumnName]))

tableMeta_writeDisposition :: Lens' TableMeta (Maybe WriteDisposition)

refComponents :: Ref a -> (Name, Name, Name)

Deconstruct a Ref into project, dataset, and table name components.

componentsRef :: (Name, Name, Name) -> Ref a