Napkin.Types.Commands

Documentation

newtype GetRelationSchema

Constructors

GetRelationSchema

Fields

_grsRelation :: Relation

newtype CheckTableExists

Constructors

CheckTableExists

Fields

_cteTable :: Ref Table

data DeleteFrom

TODO: Add in the USING clause

Constructors

DeleteFrom

Fields

_dfTable :: Ref Table

_dfWhere :: SExp

data InsertIntoQuery

Constructors

InsertIntoQuery

Fields

_iiqTable :: Ref Table

_iiqQuery :: Query

data DropTable

Drop a table

Constructors

DropTable

Fields

_dtName :: Ref Table

_dtIfExists :: Bool

_dtCascade :: Bool

data DropView

Drop a table

Constructors

DropView

Fields

_dvName :: Ref Table

_dvIfExists :: Bool

_dvCascade :: Bool

data CreateViewAs

Constructors

CreateViewAs

Fields

_cvaName :: Ref Table

_cvaQuery :: Query

data CopyTable

Constructors

CopyTable

Fields

_ctFrom :: Ref Table

_ctTo :: Ref Table

data RenameTable

Constructors

RenameTable

Fields

_rtFrom :: Ref Table

_rtTo :: Ref Table

data CreateTableAs meta

Create table based on a query with variable metadata type, such as TableMeta

Constructors

CreateTableAs

Fields

_ctaMeta :: meta

_ctaName :: Ref Table

_ctaQuery :: Query

class HasCreateTableAs bk

Operations like recreate often result in ambiguous types at application site. In general, we need to be able to infer a type for the table metadata type for each backend. Our life is complicated by the likes of Redshift TableMeta

Associated Types

type BackendTableMeta bk

createTable :: Default d => Ref Table -> Query -> CreateTableAs d

dropView :: Ref Table -> DropView

dropTable :: Ref Table -> DropTable

createToInsert :: CreateTableAs m -> InsertIntoQuery

ctaMeta :: forall meta meta. Lens (CreateTableAs meta) (CreateTableAs meta) meta meta

ctaName :: forall meta. Lens' (CreateTableAs meta) (Ref Table)

ctaQuery :: forall meta. Lens' (CreateTableAs meta) Query

cvaName :: Lens' CreateViewAs (Ref Table)

cvaQuery :: Lens' CreateViewAs Query

rtFrom :: Lens' RenameTable (Ref Table)

rtTo :: Lens' RenameTable (Ref Table)

ctFrom :: Lens' CopyTable (Ref Table)

ctTo :: Lens' CopyTable (Ref Table)

dtCascade :: Lens' DropTable Bool

dtIfExists :: Lens' DropTable Bool

dtName :: Lens' DropTable (Ref Table)

dvCascade :: Lens' DropView Bool

dvIfExists :: Lens' DropView Bool

dvName :: Lens' DropView (Ref Table)

iiqQuery :: Lens' InsertIntoQuery Query

iiqTable :: Lens' InsertIntoQuery (Ref Table)

dfTable :: Lens' DeleteFrom (Ref Table)

dfWhere :: Lens' DeleteFrom SExp

cteTable :: Iso' CheckTableExists (Ref Table)

class Drop a b | a -> b where

Methods

dropCommand :: a -> b

dropCascade :: a -> Lens' b Bool

grsRelation :: Iso' GetRelationSchema Relation

module Napkin.Types.Commands.Grant