Napkin.Untyped.Monad

Documentation

data RefStore

Supply of fresh reference names to be used in monadic contexts where you need an infinite-ish supply.

Constructors

RefStore

Fields

_refStore_feed :: [String]

_refStore_blacklist :: Set String

refStore_feed :: Lens' RefStore [String]

refStore_blacklist :: Lens' RefStore (Set String)

class HasRefStore s where

Methods

refRepo :: Lens' s RefStore

letters :: String

newRefStore :: RefStore

newRef :: (HasRefStore s, MonadState s m) => m (Ref a)

newQState :: QState

data QState

Constructors

QState

Fields

_qsVars :: RefStore

_qsQuery :: Query

qsVars :: Lens' QState RefStore

qsQuery :: Lens' QState Query

newtype Q a

Constructors

Q

Fields

_runQ :: State QState a

runQ :: Q a -> (a, Query)

mkQ :: Q a -> Query

evalQ :: Q c -> c

modQ :: Query -> Q a -> Query

Alter a pre-existing Query monadically.

asQuery :: Iso' (Q ()) Query

with_ :: Query -> Q (Ref Relation)

withQ_ :: Q a -> Q (Ref Relation)

Define a subquery using the WITH syntax. Note that this does not automatically include this table in a SELECT; you need to additionally include the resulting relation from withQ_ using the from_ operator.

withQ'_ :: Q a -> Q (Ref Relation, Query)

withQAs_ :: Ref Query -> Q a -> Q (Ref Table)

withAs_ :: Selected Query -> Q (Ref Table)

Define a WITH subquery while retaining and re-using its given name. Useful in some scenarios where you want to keep the original name in further referring to the newly introduced subquery table.

withAsUnique :: Selected Query -> Q (Ref Table)

Use unique naming scheme along with Ref Relation to construct the WITH macro's name. This is useful when ensuring unique naming while not throwing out some human-readable information.

select_ :: SExp -> Q (Ref SExp)

Smart selector that will use the variable's name if expression is a variable and assign a fresh variable name if it's something else without a name.

selectAs_ :: Selected SExp -> Q (Ref SExp)

selectsAs_ :: [Selected SExp] -> Q [Ref SExp]

selectAs_' :: Alias SExp -> Q (Maybe (Ref SExp))

Add optional aliasing to the Q.

selectsAs_' :: [Alias SExp] -> Q [Maybe (Ref SExp)]

selectsGroupsAs_ :: [Selected SExp] -> Q ()

Select a field and also include it in GROUP BY. Do this intelligently, referring to it by its given name in most cases but by its direct expression definition in case of external functions to prevent overlap with fields that may already exist.

selectsGroupsAsRef_ :: [Selected SExp] -> Q ()

Like selectsGroupsAs_ but will use the name you gave to each field in the group by statement. Some backends like BigQuery require this while others like Redshift don't like it.

selectNonEmptyGroups :: [Selected SExp] -> Q ()

Special convenience case: Select and group while demanding that none of the group values is null.

from_ :: Relation -> Q (Ref Relation)

fromIfNecessary :: Relation -> Q (Ref Relation)

Add a from clause unless the relation is already in place.

fromSelected_ :: Selected Relation -> Q (Ref Relation)

fromAlias_ :: Alias Relation -> Q (Ref Relation)

fromRef_ :: Ref Relation -> Q (Ref Relation)

Use a name directly and don't alias it one more time. | This can fail to create a proper alias if the Ref includes multiple names.

fromRel :: AsRelation r => r -> Q (Ref Relation)

fromSpecified :: Q Bool

Returns True iff the current Query already has an initial From table available.

fromOrJoin :: Alias Relation -> Either (Ref Relation -> Maybe SExp) [Ref SExp] -> Q (Ref Relation)

join_ :: JoinType -> Relation -> (Ref Relation -> Maybe SExp) -> Q (Ref Relation)

joinUsing_ :: JoinType -> Relation -> [Ref SExp] -> Q (Ref Relation)

joinRef_ :: JoinType -> Ref Relation -> (Ref Relation -> Maybe SExp) -> Q (Ref Relation)

joinSelected_ :: JoinType -> Selected Relation -> Either (Ref Relation -> Maybe SExp) [Ref SExp] -> Q (Ref Relation)

joinAlias_ :: JoinType -> Alias Relation -> Either (Ref Relation -> Maybe SExp) [Ref SExp] -> Q (Ref Relation)

joinOn_ :: JoinType -> Relation -> (Ref Relation -> SExp) -> Q (Ref Relation)

innerJoin :: Relation -> (Ref Relation -> SExp) -> Q (Ref Relation)

crossJoin :: Relation -> Q (Ref Relation)

leftJoin :: Relation -> (Ref Relation -> SExp) -> Q (Ref Relation)

rightJoin :: Relation -> (Ref Relation -> SExp) -> Q (Ref Relation)

outerJoin :: Relation -> (Ref Relation -> SExp) -> Q (Ref Relation)

having :: MonadState QState m => SExp -> m ()

havingL :: Applicative f => (Maybe SExp -> f (Maybe SExp)) -> QState -> f QState

where_ :: SExp -> Q ()

whereL :: (Maybe SExp -> Identity (Maybe SExp)) -> QState -> Identity QState

whereAdd :: MonadState QState m => SExp -> m ()

whereAddIn :: MonadState QState m => SExp -> [SExp] -> m ()

where x in y

whereAddNotIn :: MonadState QState m => SExp -> [SExp] -> m ()

where x not in y

groupBy_ :: [SExp] -> Q ()

orderBy_ :: Order -> Q ()

limit_ :: MonadState QState m => Int -> m ()

offset_ :: MonadState QState m => Int -> m ()

mdef :: a -> Iso' (Maybe a) a

fromQuery :: Q a -> Relation

Turn query into relation

scopeRelation :: SExp -> Relation -> Relation

Make a query relation that applies predicate on the given relation in a subquery.

selectEverything :: [(Ref Relation, DefinedQuery)] -> Q ()

Introspects the given list of Queryies, along with their references in the current context, and selects every named field in them within the current context, while deduplicating repeat fields (by name). Good for the common case of joining tables and selecting everything possible in them.

explicitSelectStarWith

Arguments

:: s
Object containing definitions

-> Ref Relation
Concrete reference within the context

-> Getting (Endo [Ref SExp]) s (Ref SExp)

-> Q ()

explicitSelectStar :: HasDefinition s => s -> Ref Relation -> Q ()

Instead of doing a select * that's impossible to introspect, explicitly select all fields from something that contains a definition - e.g. a Query.

selectStar :: Ref Relation -> Q (Ref Relation)

select * from x

selectNonDuplicates :: Foldable t => t (Ref SExp) -> Q ()

Given a list of references, select those that have not yet been selected in current monadic block.

mapSelectNonDuplicates :: Foldable t => t (Ref SExp) -> (Selected SExp -> Selected SExp) -> Q ()

Given a list of references, select those that have not yet been selected in current monadic block with option to modify the finalized selection.

union_ :: UnionType -> Query -> Query -> WithClauses -> Q ()