Napkin.Untyped.Ops.BigQuery

Description

Implements functions defined on https://cloud.google.com/bigquery/docs/reference/standard-sql/functions-and-operators

Documentation

windowFn :: Ref Function -> [SExp] -> WOver -> SExp

mkMod :: (a -> b) -> Maybe a -> [b]

anyValue :: SExp -> Maybe WOver -> SExp

arrayAgg :: SExp -> [FunModifier] -> SExp

arrayConcatAgg

Arguments

:: SExp

-> [Order]

-> Maybe Int
Limit

-> SExp

bqAvg :: SExp -> Either Distinctness WOver -> SExp

bitAnd :: SExp -> SExp

bitOr :: SExp -> SExp

bitXor :: SExp -> Distinctness -> SExp

bqCount :: SExp -> Distinctness -> Maybe WOver -> SExp

bqCountDistinct :: SExp -> Maybe WOver -> SExp

countIf :: SExp -> Maybe WOver -> SExp

logicalAnd :: SExp -> SExp

logicalOr :: SExp -> SExp

bqMax :: SExp -> Maybe WOver -> SExp

bqMin :: SExp -> Maybe WOver -> SExp

bqSum :: SExp -> Distinctness -> Maybe WOver -> SExp

corr :: SExp -> SExp -> SExp

Statistical Aggregate Functions

covarPop :: SExp -> SExp -> SExp

covarSamp :: SExp -> SExp -> SExp

bqStddevPop :: SExp -> Distinctness -> Maybe WOver -> SExp

bqStddevSamp :: SExp -> Distinctness -> Maybe WOver -> SExp

bqStddev :: SExp -> Distinctness -> Maybe WOver -> SExp

bqVarPop :: SExp -> Distinctness -> Maybe WOver -> SExp

bqVarSamp :: SExp -> Distinctness -> Maybe WOver -> SExp

bqVariance :: SExp -> Distinctness -> Maybe WOver -> SExp

approxCountDistinct :: SExp -> SExp

Approximate Aggregate Functions

approxQuantiles :: SExp -> Int -> Distinctness -> Maybe NullStrategy -> SExp

approxTopCount :: SExp -> Int -> SExp

approxTopSum :: SExp -> SExp -> Int -> SExp

newtype HLLPrecision

HyperLogLog++ Functions

Constructors

HLLPrecision

Fields

_unHllPrecision :: Int

hllPrecision :: Int -> HLLPrecision

hllInit :: [SExp] -> Maybe HLLPrecision -> SExp

hllMerge :: SExp -> SExp

hllMergePartial :: SExp -> SExp

hllExtract :: SExp -> SExp

rank :: WOver -> SExp

Analytic (Window) Functions

denseRank :: WOver -> SExp

percentRank :: WOver -> SExp

rowNumber :: WOver -> SExp

cumeDist :: WOver -> SExp

ntile :: Int -> WOver -> SExp

firstValue :: SExp -> WOver -> Maybe NullStrategy -> SExp

lastValue :: SExp -> WOver -> Maybe NullStrategy -> SExp

nthValue :: SExp -> Int -> WOver -> Maybe NullStrategy -> SExp

lead :: SExp -> Int -> Maybe SExp -> WOver -> SExp

lag :: SExp -> Int -> Maybe SExp -> WOver -> SExp

percentileCont :: SExp -> Double -> [SExp] -> Maybe NullStrategy -> SExp

percentileDisc

Arguments

:: SExp
Value being percentiled

-> Double
Percentile boundary

-> [SExp]
Partition clause

-> Maybe NullStrategy

-> SExp

bitCount :: SExp -> SExp

Bit Functions

type UFunc = SExp -> SExp

Math Functions

type BinFunc = SExp -> SExp -> SExp

type NulFunc = SExp

type MulFunc = [SExp] -> SExp

isInf :: UFunc

isNan :: UFunc

ieeeDivide :: BinFunc

rand :: NulFunc

sqrt :: UFunc

pow :: BinFunc

power :: BinFunc

exp :: UFunc

ln :: UFunc

log :: BinFunc

log10 :: UFunc

greatest :: MulFunc

least :: MulFunc

div :: BinFunc

safeDiv :: BinFunc

mod :: BinFunc

round :: UFunc

roundN :: SExp -> Int -> SExp

trunc :: UFunc

truncN :: Val a => SExp -> a -> SExp

ceil :: UFunc

floor :: UFunc

cos :: UFunc

cosh :: UFunc

acos :: UFunc

acosh :: UFunc

sin :: UFunc

sinh :: UFunc

asin :: UFunc

asinh :: UFunc

tan :: UFunc

tanh :: UFunc

atan :: UFunc

atanh :: UFunc

atan2 :: BinFunc

farmFingerprint :: UFunc

Hash Functions

md5 :: UFunc

sha1 :: UFunc

sha256 :: UFunc

sha512 :: UFunc

byteLength :: UFunc

String Functions

charLength :: UFunc

codePointsToBytes :: UFunc

codePointsToString :: UFunc

concat :: MulFunc

endsWith :: BinFunc

format :: String -> [SExp] -> SExp

fromBase64 :: UFunc

length :: UFunc

lpad :: SExp -> Int -> Maybe String -> SExp

lower :: UFunc

ltrim :: SExp -> Maybe String -> SExp

regexpContains :: SExp -> String -> SExp

regexpExtract :: SExp -> String -> SExp

regexpExtractAll :: SExp -> String -> SExp

regexpReplace :: SExp -> String -> String -> SExp

replace :: SExp -> String -> String -> SExp

repeat :: Val a => SExp -> a -> SExp

reverse :: UFunc

rpad :: SExp -> Int -> Maybe String -> SExp

rtrim :: SExp -> Maybe String -> SExp

safeConvertBytesToString :: UFunc

safeCast :: SExp -> Type -> SExp

split :: SExp -> Text -> SExp

startsWith :: SExp -> SExp -> SExp

strPos :: SExp -> SExp -> SExp

substr :: SExp -> Int -> Maybe Int -> SExp

toBase64 :: UFunc

toCodePoints :: UFunc

toHex :: UFunc

trim :: SExp -> Maybe String -> SExp

upper :: UFunc

data JSONPath

JSON Functions

Constructors

JSONPath_Key String

JSONPath_Index !Int

jsonPath_key :: String -> JSONPath

jsonPath_index :: Int -> JSONPath

renderJSONPath :: [JSONPath] -> String

jsonExtract :: SExp -> [JSONPath] -> SExp

jsonExtractScalar :: SExp -> [JSONPath] -> SExp

toJsonString

Arguments

:: SExp

-> Maybe Bool
Pretty print

-> SExp

arrayConcat :: NonEmpty SExp -> SExp

Array Functions

arrayLength :: SExp -> SExp

arrayToString

Arguments

:: SExp

-> String
Delimiter

-> Maybe String
Null text

-> SExp

generateArray

Arguments

:: SExp
Start exp

-> SExp
End exp

-> Maybe SExp
Step exp

-> SExp

generateDateArray :: SExp -> SExp -> Maybe IntInterval -> SExp

arrayReverse :: SExp -> SExp

offset :: Int -> SExp -> SExp

ordinal :: Int -> SExp -> SExp

currentDate :: Maybe String -> SExp

Date Functions

dateYMD

Arguments

:: SExp
Year

-> SExp
Month

-> SExp
Day

-> SExp

dateTimestamp

Arguments

:: SExp
Timestamp exp

-> Maybe SExp
Timezone exp

-> SExp

dateAdd

Arguments

:: SExp
Date expression

-> IntInterval

-> SExp

dateSub

Arguments

:: SExp
Date expression

-> IntInterval

-> SExp

dateDiff :: DatePart -> SExp -> SExp -> SExp

dateTrunc :: DatePart -> SExp -> SExp

dateFromUnixDate

Arguments

:: SExp
Int64 expr

-> SExp

formatDate

Arguments

:: SExp
Format string

-> SExp
Date expr

-> SExp

parseDate

Arguments

:: SExp
Format string

-> SExp
Date expr

-> SExp

unixDate

Arguments

:: SExp
Date expr

-> SExp

currentDatetime

Arguments

:: Maybe SExp
Timezone

-> SExp

DateTime Functions

dateTimeYMDHMS

Arguments

:: SExp
Year

-> SExp
Month

-> SExp
Day

-> SExp
Hour

-> SExp
Minute

-> SExp
Second

-> SExp

dateTimeDateAndTime

Arguments

:: SExp
Date expression

-> SExp
Time expression

-> SExp

dateTimeTimestamp

Arguments

:: SExp
Timestamp expression

-> Maybe SExp
Timezone expression

-> SExp

dateTimeAdd

Arguments

:: SExp
Datetime expression

-> IntInterval

-> SExp

dateTimeSub

Arguments

:: SExp
Datetime expression

-> IntInterval

-> SExp

dateTimeDiff :: DatePart -> SExp -> SExp -> SExp

dateTimeTrunc :: DatePart -> SExp -> SExp

formatDateTime

Arguments

:: SExp
Format string

-> SExp
Datetime expression

-> SExp

parseDateTime

Arguments

:: SExp
Format string

-> SExp
Datetime expression

-> SExp

currentTime :: SExp

Time Functions

timeHMS

Arguments

:: SExp
Hour

-> SExp
Minute

-> SExp
Second

-> SExp

timeTS

Arguments

:: SExp
Timestamp

-> Maybe SExp
Time zone

-> SExp

timeAdd

Arguments

:: SExp
Time expression

-> IntInterval

-> SExp

timeSub

Arguments

:: SExp
Time expression

-> IntInterval

-> SExp

timeDiff :: DatePart -> SExp -> SExp -> SExp

timeTrunc :: DatePart -> SExp -> SExp

bqFormatTime

Arguments

:: SExp
Format string

-> SExp
Time object

-> SExp

bqParseTime

Arguments

:: SExp
Format string

-> SExp
Time object

-> SExp

currentTimestamp :: SExp

Timestamp Functions

extract

Arguments

:: DatePart

-> SExp
Timestamp expression

-> Maybe SExp
Time zone spec

-> SExp

string

Arguments

:: SExp
Timestamp expression

-> Maybe SExp
Time zone

-> SExp

timestamp

Arguments

:: SExp
string, date, or datetime expression

-> Maybe SExp
Time zone

-> SExp

timestampAdd :: SExp -> IntInterval -> SExp

timestampSub :: SExp -> IntInterval -> SExp

timestampDiff :: DatePart -> SExp -> SExp -> SExp

timestampTrunc

Arguments

:: DatePart

-> SExp
Timestamp expression

-> Maybe SExp
Time zone

-> SExp

formatTimestamp

Arguments

:: SExp
Format string

-> SExp
Timestamp

-> Maybe SExp
Time zone

-> SExp

parseTimestamp

Arguments

:: SExp
Format string

-> SExp
Timestamp

-> Maybe SExp
Time zone

-> SExp

timestampSeconds

Arguments

:: SExp
Int64 expression

-> SExp

timestampMillis

Arguments

:: SExp
Int64 expression

-> SExp

timestampMicros

Arguments

:: SExp
Int64 expression

-> SExp

unixSeconds

Arguments

:: SExp
Timestamp expression

-> SExp

unixMillis

Arguments

:: SExp
Timestamp expression

-> SExp

unixMicros

Arguments

:: SExp
Timestamp expression

-> SExp

sessionUser :: SExp

Security Functions

netIPFromString

Arguments

:: SExp
Addr string

-> SExp

Net Functions

netSafeIPFromString

Arguments

:: SExp
Addr string

-> SExp

Like netIPFromString but returns NULL rather than throwing on error.

netIPToString

Arguments

:: SExp
Addr binary

-> SExp

netIPNetMask

Arguments

:: SExp
Num output bytes

-> SExp
Prefix length

-> SExp

netIPTrunc

Arguments

:: SExp
Addr binary

-> SExp
Prefix length

-> SExp

netIPV4FromInt64

Arguments

:: SExp

-> SExp
Addr binary

netIPV4ToInt64

Arguments

:: SExp
Addr binary

-> SExp

netHost

Arguments

:: SExp
URL

-> SExp

netPublicSuffix

Arguments

:: SExp
URL

-> SExp

netRegDomain

Arguments

:: SExp
URL

-> SExp

coalesce :: [SExp] -> SExp

Conditional Expressions

ifnull

Arguments

:: SExp
Expressoin

-> SExp
Null result

-> SExp

nullif

Arguments

:: SExp
Expression

-> SExp
Expression to match

-> SExp

like :: SExp -> SExp -> SExp

Comparison operators

between :: SExp -> SExp -> SExp -> SExp

scope :: SExp -> SExp -> SExp

Higher level combinators

scoped :: SExp -> SExp -> SExp

inferAgg :: SExp -> AggLevel

unitBoundary :: SExp -> [SExp]

modifyUnitBoundary :: (SExp -> SExp) -> SExp -> SExp

bigQueryMeta :: BackendFunctionMeta

bigQueryAggs :: [(Ref Function, SExp)]

textInfix :: SExp -> SExp -> SExp

Is text a contained in text b?

nullIfEmpty :: SExp -> SExp

Replace empty string with NULL

seekBackIgnoreNulls :: SExp -> Int -> WOver -> SExp

Use nthValue under the hood with IgnoreNulls, set window frame to start at current row (so nth-value can index) and flip the ordering in given OVER clause so that it semantically behaves like lag but with ignore nulls functionality.

It will start with current row, so setting to 1 will just return current row's value.

seekForwardIgnoreNulls :: SExp -> Int -> WOver -> SExp

Same deal as lagIgnorenulls but doesn't need to flip the ordering in OVER clause as lead and nthValue both naturally operate in ascending order.

lagIgnoreNulls :: SExp -> Int -> WOver -> SExp

BigQuery can't ignore nulls while lagging/leading. Use the seekBackignorenulls machinery intelligently to mimic a proper lag with ignore nulls.

leadIgnoreNulls :: SExp -> Int -> WOver -> SExp

denseAvg :: SExp -> SExp

weightedAvg :: SExp -> SExp -> SExp

nzAvg :: SExp -> SExp

percentDiff :: SExp -> SExp -> SExp

percent difference of a over b with typecast to Double

data SubQueryTransformState

Constructors

SubQueryTransformState

Fields

_subQueryTransformState_refs :: RefStore

_subQueryTransformState_mapping :: Map SExp (Ref SExp)

subQueryTransformState_refs :: Lens' SubQueryTransformState RefStore

subQueryTransformState_mapping :: Lens' SubQueryTransformState (Map SExp (Ref SExp))

subqueryTransform :: Query -> Query

humanAge :: SExp -> SExp -> SExp

Calculates the year-based human age given a "current date" and a date in the past. Returns float-based age. DATE_DIFF(CURRENT_DATE, cast('1990-01-01' as date), YEAR) + IF(EXTRACT(DAYOFYEAR FROM CURRENT_DATE) < EXTRACT(DAYOFYEAR FROM cast('1990-01-01' as date)), -1, 0)