Napkin.Untyped.Ops.Redshift

Documentation

toChar :: SExp -> Text -> SExp

Type casts

toDate :: Text -> Text -> SExp

toNumber :: Text -> Text -> SExp

stddev_samp :: SExp -> SExp

Aggregate functions

stddev :: SExp -> SExp

stddev_pop :: SExp -> SExp

acos :: SExp -> SExp

Math functions

asin :: SExp -> SExp

atan :: SExp -> SExp

atan2 :: SExp -> SExp

cbrt :: SExp -> SExp

ceil :: SExp -> SExp

checksum :: SExp -> SExp

cos :: SExp -> SExp

cot :: SExp -> SExp

degrees :: SExp -> SExp

dexp :: SExp -> SExp

dlog1 :: SExp -> SExp

dlog10 :: SExp -> SExp

exp :: SExp -> SExp

floor :: SExp -> SExp

ln :: SExp -> SExp

log :: SExp -> SExp

mod :: SExp -> SExp -> SExp

pi :: SExp

power :: SExp -> SExp

radians :: SExp -> SExp

random :: SExp

round :: SExp -> SExp

sin :: SExp -> SExp

sqrt :: SExp -> SExp

tan :: SExp -> SExp

trunc :: SExp -> Maybe Int -> SExp

greatest :: [SExp] -> SExp

Conditionals

least :: [SExp] -> SExp

coalesce :: [SExp] -> SExp

nvl :: [SExp] -> SExp

nvl2 :: SExp -> SExp -> SExp -> SExp

nullif :: SExp -> SExp -> SExp

decode :: SExp -> [(SExp, SExp)] -> Maybe SExp -> SExp

data CaseSensitivity

Equality

Constructors

CaseSensitive

CaseInSensitive

like :: CaseSensitivity -> SExp -> SExp -> SExp

(||.) :: SExp -> SExp -> SExp

String ops

Concatenate two strings using ||

bpcharcmp :: SExp -> SExp -> SExp

btrim :: SExp -> Maybe String -> SExp

charindex :: SExp -> SExp -> SExp

chr :: SExp -> SExp

crc32 :: SExp -> SExp

funcSha1 :: SExp -> SExp

initCap :: SExp -> SExp

left :: SExp -> SExp -> SExp

right :: SExp -> SExp -> SExp

len :: SExp -> SExp

lower :: SExp -> SExp

lpad :: SExp -> SExp -> Maybe SExp -> SExp

rpad :: SExp -> SExp -> Maybe SExp -> SExp

ltrim :: SExp -> SExp -> SExp

md5 :: SExp -> SExp

position :: SExp -> SExp -> SExp

quoteIdent :: SExp -> SExp

quoteLiteral :: SExp -> SExp

regexpCount :: SExp -> SExp -> Maybe SExp -> SExp

regexpInstr :: SExp -> SExp -> Maybe SExp -> SExp

regexpReplace :: SExp -> SExp -> SExp -> Maybe SExp -> SExp

regexpSubstr :: SExp -> SExp -> Maybe SExp -> SExp

repeat :: SExp -> SExp -> SExp

replace :: SExp -> SExp -> SExp -> SExp

reverse :: SExp -> SExp

rtrim :: SExp -> SExp -> SExp

splitPart :: SExp -> SExp -> SExp -> SExp

strtol :: SExp -> SExp -> SExp

substring :: SExp -> Int -> Int -> SExp

toHex :: SExp -> SExp

translate :: SExp -> SExp -> SExp -> SExp

trim :: Maybe SExp -> SExp -> SExp

upper :: SExp -> SExp

addMonths :: SExp -> SExp -> SExp

Date functions

convertTimeZone :: SExp -> SExp -> SExp -> SExp

current_date :: SExp

timeOfDay :: SExp

dateCmp :: SExp -> SExp -> SExp

dateCmpTimestamp :: SExp -> SExp -> SExp

datePartYear :: SExp -> SExp

dateAdd :: DatePart -> SExp -> SExp -> SExp

dateDiff :: DatePart -> SExp -> SExp -> SExp

datePart :: DatePart -> SExp -> SExp

dateTrunc :: DatePart -> SExp -> SExp

extract :: DatePart -> SExp -> SExp

getDate :: SExp

intervalCmp :: SExp -> SExp -> SExp

lastDay :: SExp -> SExp

monthsBetween :: SExp -> SExp -> SExp

nextDay :: SExp -> SExp -> SExp

sysdate :: SExp

timestampCmp :: SExp -> SExp -> SExp

timestampCmpDate :: SExp -> SExp -> SExp

wCount :: SExp -> WOver -> SExp

Window functions

wavg :: SExp -> WOver -> SExp

wsum :: SExp -> WOver -> SExp

wmin :: SExp -> WOver -> SExp

wmax :: SExp -> WOver -> SExp

wmedian :: SExp -> Partition -> SExp

ntile :: SExp -> Partition -> Order -> SExp

percentRank :: SExp -> Partition -> Order -> SExp

cumeDist :: WOver -> SExp

denseRank :: WOver -> SExp

firstValue :: SExp -> Maybe NullStrategy -> WOver -> SExp

lastValue :: SExp -> Maybe NullStrategy -> WOver -> SExp

lag :: SExp -> Int -> WOver -> Maybe NullStrategy -> SExp

lead :: SExp -> Int -> WOver -> Maybe NullStrategy -> SExp

nthValue :: SExp -> Int -> WOver -> Maybe NullStrategy -> SExp

listAgg :: SExp -> Maybe SExp -> Order -> Partition -> Maybe NullStrategy -> SExp

percentileCont :: Double -> Order -> [SExp] -> SExp

percentileDisc :: Double -> Order -> [SExp] -> SExp

rank :: WOver -> SExp

ratioToReport :: SExp -> [SExp] -> SExp

rowNumber :: WOver -> SExp

wstddev :: SExp -> WOver -> SExp

wstddevSamp :: SExp -> WOver -> SExp

wstddevPop :: SExp -> WOver -> SExp

variance :: SExp -> WOver -> SExp

varSample :: SExp -> WOver -> SExp

varPop :: SExp -> WOver -> SExp

denseAvg :: SExp -> SExp

Higher level combinators

weightedAvg :: SExp -> SExp -> SExp

nzAvg :: SExp -> SExp

safeDiv :: SExp -> SExp -> SExp

percentDiff :: SExp -> SExp -> SExp

percent difference of a over b with typecast to Double

scope :: SExp -> SExp -> SExp

scoped :: SExp -> SExp -> SExp

redshiftAggs :: [(Ref Function, SExp)]

textInfix :: SExp -> SExp -> SExp

Is text a contained in text b?

generateSeries :: SExp -> SExp -> SExp -> SExp

Other functions

dateDiffSafe :: DatePart -> SExp -> SExp -> SExp

Special Cases

Redshift seems to have a bug where it will overflow when asking day ranges that become too negative, typically due to faulty columns that have bogus dates like "Nov 1 2065" when referring to values that should be around now.

Keep in mind we do not know exactly how the bug occurs; this is a guess based on what we've seen in a one-off edge condition.

dateDiffNonNeg :: DatePart -> SExp -> SExp -> SExp

nilBrokenDate :: SExp -> SExp

nullIfEmpty :: SExp -> SExp

Replace empty string with NULL