Napkin.Untyped.Utils

Documentation

combineTablesLeft

Arguments

:: DefinedRelation

-> [(DefinedRelation, Ref Relation -> Ref Relation -> SExp)]
Table definitions and primary keys each shares with the first table in list

-> Q ()

Combine multiple tables/queries that were defined in napkin that share a common set of primary keys via a chain of left joins.

columnsToRows

Arguments

:: DefinedRelation

-> [String]
Groups, keep these intact on each row

-> Query

Transfroms columns into separate rows, per given partition

unionTables :: [DefinedRelation] -> Query

Combine all tables into one assuming if the field names are the same, their types will just work out. Fill missing columns with NULL on a per-table basis.

unionWithSchema :: [(Relation, Fields Field)] -> Query

Unify schemas and join given relations under a single UNION query.

unifySchemas :: [(Relation, Fields Field)] -> [Query]

Given relations' and their known schema, whether it's a Query or a table reference, union them in a way that works out the missing columns in each and even unifies mis-matching types to a common denominator.

castToField :: Field -> SExp -> SExp

Cast expression bsaed on a desired type from the sql-schema library.

filterFields :: (Ref t -> Bool) -> Query -> Query

Only keep selected fields in Query that pass the predicate.

whitelistFields :: Foldable t1 => t1 Regex -> Ref t -> Bool

Keep any field that matches any of the given regexes. Can use this to construct the predicate in filterFields

longToWideFormat :: [Alias SExp] -> [(SExp, SExp, Ref SExp)] -> Query -> Query

Convert a "long" query to "wide" by aggregating/matching over certain SExps.

longToWideFormatSimple :: [Alias SExp] -> [(SExp, SExp)] -> Ref SExp -> SExp -> Query -> Query

A long-to-wide query transformer which matches using equality