Napkin.Utils

Documentation

for :: [a] -> (a -> b) -> [b]

anyRight :: Foldable t => t (Either a b) -> Bool

whenJust :: Applicative f => Maybe t -> (t -> f ()) -> f ()

collect

Arguments

:: (Ord k, Foldable t)

=> (a -> Maybe k)
key in collected map

-> (a -> v)
value in collected map

-> (v -> v -> v)
collapse function for value

-> t a
something I can traverse, e.g. [a]

-> Map k v

makeLensesTrail :: Name -> Q [Dec]

nameRule :: a -> b -> Name -> [DefName]

echo :: Katip m => Severity -> LogStr -> m ()

echo' :: (Katip m, LogItem a) => a -> Severity -> LogStr -> m ()

timeLog :: Katip m => String -> m b -> m b

time :: MonadIO m => m a -> m (Double, a)

Measure how long action took, in seconds along with its result

time_ :: MonadIO m => m a -> m Double

Just measure how long action takes, discard its result

enumUniverse :: forall a. (Bounded a, Enum a) => [a]

getTime :: IO Double

getTempTableName :: MonadIO m => m (Ref b)

randomToken :: MonadIO m => m String

urlencode :: (StringConv ByteString b, StringConv a ByteString) => a -> Tagged "urlencoded" b

Encode most string-like types into a tagged, urlencoded form. Tagging helps ensure that it is indeed encoded and won't be double-encoded

urlencodedProxy :: Proxy "urlencoded"

Can be used with Tagged to indicate any textual value is urlencoded. tagWith urlencodedProxy to wrap, proxy urlencodedProxy to unwrap.

prettyBytes :: Int64 -> String

prettySeconds :: Double -> String

replaceChar :: Char -> Char -> String -> String